% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R
\name{checkRloadestStatus}
\alias{checkRloadestStatus}
\title{Actions & checks for when the user [indirectly] calls loadReg()}
\usage{
checkRloadestStatus()
}
\description{
Require that the user has loaded rloadest. Also give the user a message about
rloadest if appropriate.
}
\keyword{internal}

