% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETDaily}
\alias{convertToEGRETDaily}
\title{Convert estimation and load prediction data into the EGRET Daily data.frame}
\usage{
convertToEGRETDaily(estdat, preds, meta)
}
\arguments{
\item{estdat}{data.frame of estimation data}

\item{preds}{data.frame of load predictions}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}
}
\description{
Convert estimation and load prediction data into the EGRET Daily data.frame
}

