% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETInfo}
\alias{convertToEGRETInfo}
\title{Convert a loadflex metadata object into the EGRET INFO dataframe.}
\usage{
convertToEGRETInfo(meta)
}
\arguments{
\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}
}
\description{
Convert a loadflex metadata object into the EGRET INFO dataframe.
}

