% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETSample}
\alias{convertToEGRETSample}
\title{Convert the interpolation data.frame into the EGRET Sample dataframe.}
\usage{
convertToEGRETSample(intdat, meta)
}
\arguments{
\item{intdat}{data.frame of interpolation data}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}
}
\description{
Convert the interpolation data.frame into the EGRET Sample dataframe.
}

