% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{convertUnits}
\alias{convertUnits}
\title{Produce conversion factor to multiply old.units by to get new.units}
\usage{
convertUnits(old.units, new.units, attach.units = FALSE)
}
\arguments{
\item{old.units}{character. The units to convert from.}

\item{new.units}{character. The units to convert to.}

\item{attach.units}{logical. Should units be attached to the conversion 
factor?}
}
\value{
a conversion factor that can be multiplied with data in the old.units
  to achieve data in the new.units
}
\description{
Converts units; units can be arbitrarily complex as long as every dimension 
of each unit is present in unit.conversions
}
\examples{
loadflex:::convertUnits('mg/L', 'kg/m^3')
loadflex:::convertUnits('kg d^-1', 'kg/yr')
loadflex:::convertUnits('kg/yr', 'kg/d', attach.units = TRUE)
\dontrun{
loadflex:::convertUnits('mg/L', 'm^3/d') # error: dimensions must match
loadflex:::convertUnits(unitbundle('ft^3 g L^-1 s^-1'), 'kg/d') # error: too complicated
}
}
