% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{correlations-1D}
\alias{correlations-1D}
\alias{rhoEqualDates}
\alias{rho1DayBand}
\title{Get the assumed correlation between residuals or predictions at pairs of 
dates.}
\usage{
rhoEqualDates(date1, date2)

rho1DayBand(date1, date2)
}
\arguments{
\item{date1}{The date of the first prediction}

\item{date2}{The date[s] of the second prediction[s].}
}
\value{
A correlation coefficient or a vector of coefficients corresponding 
  to the pair or pairs in the rows of \code{cbind(date1, date2)}.
}
\description{
These 1D correlation functions are made available to help users explore and 
understand the assumptions that may be made about the correlation structure 
of residuals or predictions. It is possible to use these functions directly 
for aggregation by calling \code{\link{aggregateSolute}} with the 
\code{cormat.function} argument set to either 
\code{\link{getCormatCustom}(rhoEqualDates)} or 
\code{\link{getCormatCustom}(rho1DayBand)}. However, the equivalent but more 
efficient options are to set \code{cormat.function} to 
\code{\link{cormatEqualDates}} or \code{\link{cormat1DayBand}}, respectively,
because those \code{cormatXXX} functions produce the full correlation matrix 
without costly intermediate steps.
}
\details{
\code{rhoEqualDates} embodies the standard LOADEST/rloadest assumption about 
correlation of residuals for the purpose of aggregation. Specifically, if two
point predictions are on the same calendar date, then rho (the correlation of
their errors) is set to 1, and otherwise rho = 0.

\code{rho1DayBand} returns 1 for pairs of date-times that are within 1 day
(86400 seconds) of each other, or 0 for pairs that are not. This is a comparable
but smoother assumption than \code{rhoEqualDates}.
}
