% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{correlations-2D}
\alias{cormat1DayBand}
\alias{cormatDiagonal}
\alias{cormatEqualDates}
\alias{correlations-2D}
\title{Functions that each produce an autocorrelation matrix with a specified 
pattern.}
\usage{
cormatEqualDates(dates)

cormat1DayBand(dates)

cormatDiagonal(dates)
}
\arguments{
\item{dates}{date-times (as Date, POSIXct, chron, etc.) from which the 
autocorrelation matrix should be produced.}
}
\value{
A matrix of autocorrelation coefficients for all possible pairs of 
  date-times in \code{dates}.
}
\description{
Accepts a vector of date-times or dates and returns a correlation matrix 
describing the assumed correlations between all possible pairs of those 
dates.

\code{cormatEqualDates} formalizes the assumption used by LOADEST and
rloadest: if two date-times are on the same calendar date, the correlation
(rho) is 1. Otherwise, rho=0.

\code{cormat1DayBand} produces a matrix containing 1s within a 1-day band 
(within 24 hours preceding or 24 hours following an observation) and 0s 
elsewhere. This is a special case of the functions that can be generated by 
\code{\link{getCormatTaoBand}()}.

\code{cormatDiagonal} produces a matrix containing 1s ONLY along the diagonal
- i.e., no covariance among predictions. This function is rarely suitable for aggregation with uncertainty estimation,
but it's faster than the alternatives.
}

