% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{correlations}
\alias{correlations}
\title{Correlation functions in \pkg{loadflex}}
\description{
Correlations of residuals or prediction errors are important at several
points in the process of estimating solute concentrations or fluxes. Directly
estimating these correlations is difficult, though sometimes possible. The
functions listed here exist to help the user in (1) estimating correlations
from data, and (2) asserting correlation structures when empirical estimates
are weak or unavailable.

\subsection{1D Correlation Functions}{

These are functions that produce one or more correlation coefficients from pairs of dates.

\itemize{

\item \code{\link{rhoEqualDates}}

\item \code{\link{rho1DayBand}}

}

}

\subsection{1D Correlation Function Generators}{

These are functions that produce functions that produce one or more correlation coefficients.

\itemize{

\item \code{\link{getRhoFirstOrderFun}}

}

}

\subsection{2D Correlation Functions}{

These are functions that produce a 2D correlation matrix from a vector of dates.

\itemize{

\item \code{\link{cormatEqualDates}}

\item \code{\link{cormat1DayBand}}

}

}

\subsection{2D Correlation Function Generators}{

These are functions that produce functions that produce 2D correlation matrices.

\itemize{

\item \code{\link{getCormatCustom}}

\item \code{\link{getCormatTaoBand}}

\item \code{\link{getCormatFirstOrder}}

}

}
}

