% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{dimInfo}
\alias{dimInfo}
\title{Return a data.frame describing each dimension of the units (1 row per dimension)}
\usage{
dimInfo(unitstr)
}
\arguments{
\item{unitstr}{A single string representing one set of units dimension}
}
\description{
Return a data.frame describing each dimension of the units (1 row per dimension)
}
\examples{
loadflex:::dimInfo('kg') # 'mg'
loadflex:::dimInfo('ha') # 'km^2'
loadflex:::dimInfo('kg d^-1') # NA
loadflex:::dimInfo('m^3') # NA
loadflex:::dimInfo('kk') # NA
}
