% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{eg_loadflex}
\alias{eg_loadflex}
\alias{lamprey_discharge}
\alias{lamprey_nitrate}
\alias{eg_fitdat}
\alias{eg_estdat}
\alias{eg_metadata}
\alias{eg_loadInterp}
\alias{eg_loadLm}
\alias{eg_loadReg2}
\alias{eg_loadComp}
\title{Example datasets and objects for the \pkg{loadflex} package}
\description{
These datasets and pre-created objects are provided for 
  exploring and testing the \pkg{loadflex} package.
}
\section{Datasets}{

  
  \describe{
  
  \item{\code{eg_fitdat}}{Example dataset for fitting (calibrating) a model. 
  These data are a lightweight subset of \code{lamprey_nitrate}, below.}
  
  \item{\code{eg_estdat}}{Example dataset for generating predictions from a 
  fitted model. These data are a lightweight subset of 
  \code{lamprey_discharge}, below.}
  
  \item{\code{lamprey_discharge}}{Discharge data for the Lamprey River from 
  10/1/1999 to 11/16/2014. Discharge in CFS measured every  15 minutes and 
  collected by the US Geological Survey, site 01073500, waterdata.usgs.gov. 
  The Lamprey River is an 81-km river flowing through southeastern New 
  Hampshire. Its 548-km2 watershed empties into the Great Bay estuary.}
  
  \item{\code{lamprey_nitrate}}{Nitrate data for the Lamprey River from 
  10/1/1999 to 11/16/2014. Nitrate is in mg/L and is collected weekly. The 
  Lamprey River is an 81-km river flowing through southeastern New Hampshire.
  Its 548-km2 watershed empties into the Great Bay estuary. Nitrate 
  concentrations have been monitored with weekly and event-based grab samples
  at Packers Falls on the Lamprey since 10 September 1999 and are measured by
  SmartChem discrete analyzer (Westco, Brookfield, CT).}
  
  }
}

\section{Objects}{

  
  \describe{
  
  \item{\code{eg_metadata}}{Example metadata object.}
  
  \item{\code{eg_loadInterp}}{Example interpolation model object.}
  
  \item{\code{eg_loadLm}}{Example linear regression model object.}
  
  \item{\code{eg_loadReg2}}{Example model object containing an inner 
  \pkg{rloadest} model.}
  
  \item{\code{eg_loadComp}}{Example composite method model object.}
  
  }
}

