% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{flowUnitsConversion}
\alias{flowUnitsConversion}
\title{Get a conversion factor to convert between flow units}
\usage{
flowUnitsConversion(old.units, new.units, attach.units = FALSE)
}
\arguments{
\item{old.units}{character. The current units of flow.}

\item{new.units}{character. The desired units of flow.}

\item{attach.units}{logical. If TRUE, the conversion factor is returned with 
units attached.}
}
\value{
numeric conversion factor, to be multiplied by values in old units to
  determine the values in new units.
}
\description{
Get a conversion factor to convert between flow units
}
\examples{
flowUnitsConversion(old.units='cfs', new.units='cms')
flowUnitsConversion(old.units='m^3 s^-1', new.units='ft^3 s^-1')
flowUnitsConversion(old.units='m^3 s^-1', new.units='ft^3 s^-1', attach.units=TRUE)
}

