% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leverage.R
\name{fluxInfluence}
\alias{fluxInfluence}
\title{Covariance ratio}
\usage{
fluxInfluence(fit.function, obs.calib, obs.adjust, data.to.predict,
  correction.method = "linear", replace.negatives = -1,
  aggregation.interval = "unit", influence.on = c("both", "regression",
  "adjustment"))
}
\arguments{
\item{fit.function}{function(...) that
takes a training predictors obs.calib X.calib, an observation Y.calib, a
prediction matrix X.est, and a character or factor vector P.est that
identifies the Period of each row of X.est. The function returns the
aggregated flux predictions Fhat for the observations X.est given a model
(in our case, usually a composite-method model) fitted to X.calib and
Y.calib.}

\item{obs.calib}{a vector of concentration observations}

\item{obs.adjust}{of concentrations less the observation(s) to be tested for influence.}

\item{data.to.predict}{a vector of prediction predictors}

\item{correction.method}{default is linear, passed to predictSolute}

\item{replace.negatives}{passed to predictSolute,}

\item{aggregation.interval}{for the predictions}

\item{influence.on}{calculate influence on regression, adjustment, or both}
}
\value{
A vector of the same length as m with an influence measure for each
  row index (of X and/or Y) in m.
}
\description{
CR - det(s_i^2 * (X_i' %*% X_i)^-1) / det(s^2 * (X' %*% X)^-1)
}
\details{
What does this tell us about the predictions, as opposed to the coefficients?

I don't think this equation works; It's based on an assumption that the
predictions are directly related to the Xes, and it also requires an estimate
of the standard eror. Both of these are problematic to me at this point.
DFFITS # conceptually identical to cook's distance

DFFITS = (yhat_i - yhat_iI) / (s_I*sqrt(h_ii))

where I is (i) means "without observation i", yhat_i and yhat_iI are the
prediction for point i with and without point i included in the model, s_i is
the standard error estimated without point i, and hii is the leverage for the
point. (http://en.wikipedia.org/wiki/DFFITS)

h_ii is 1. What's s_I? That could be a problem.
Influence

I would like to find a reference to justify this particular approach, but I
think the basic idea is about right: the influence of a particular
observation on a flux estimate for a given period is the difference between
the flux estimate with that observation and the flux estimate without.
}

