% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit.conversions.R
\docType{data}
\name{freeform.unit.translations}
\alias{freeform.unit.translations}
\title{Defines the units that are permissible to pass to load models}
\format{A data.frame with character columns "new" and "old"}
\description{
Contains a dictionary with which we will translate a range of possibilities
into a smaller, more uniform set of units
}
\keyword{data}
\keyword{units}

