% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generalizedLeverage}
\alias{generalizedLeverage}
\title{Find the generalized leverage of observation m when perturbed by distance b.}
\usage{
generalizedLeverage(b, m, fit.and.predict.function, X, Y)
}
\arguments{
\item{b}{numeric value or vector. The distance(s) by which observation m
should be perturbed}

\item{m}{numeric value. The index of the observation in Y that should be
perturbed.}

\item{fit.and.predict.function}{function(X, Y) that takes a predictors matrix X and
an observation Y and returns the predictions Yhat for the observations X
given a model (in our case, usually a composite-method model) fitted to X and Y.}

\item{X}{an (n,p) matrix of predictors}

\item{Y}{an n-length vector of observations}
}
\value{
return the leverage of m on X
}
\description{
Generalized leverage: G(b;m) = (1/b)*(Yhatmb - Yhat), i.e., the difference in
fitted values between the models without and with a perturbation of size b on
the mth observation, divided by b (as from St. Laurent and Cook, 1992)
}
\details{
You can have high leverage without having high influence; influence =
leverage * discrepancy.
}

