% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{generateUnitsData}
\alias{generateUnitsData}
\title{Generate units conversion & parsing data}
\usage{
generateUnitsData()
}
\description{
Should only need to be run once on a change in the function contents.
Generates the units data to be saved in data/valid.metadata.units, 
data/freeform.unit.translations, data/unit.conversions, and (all three 
combined) R/sysdata.rda. These are then saved with the package.
}
\keyword{data}
\keyword{internal}
\keyword{units}

