% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolations.R
\name{genericDistanceWeightedInterpolation}
\alias{genericDistanceWeightedInterpolation}
\title{A parameterizable distance-weighted interpolation function.}
\usage{
genericDistanceWeightedInterpolation(
  dates.in,
  y.in,
  dates.out,
  inv.dist.fun = function(a, b) {
     1/((a - b)^2)
 }
)
}
\arguments{
\item{dates.in}{A numeric vector desribing the dates for each of the values 
in \code{y.in}. Dates are represented as the number of seconds since 1970.}

\item{y.in}{A vector of values (typically fluxes or concentrations) to 
interpolate among.}

\item{dates.out}{A numeric vector of dates for which the corresponding output
values are to be produced. Dates are represented as the number of seconds 
since 1970.}

\item{inv.dist.fun}{A function to calculate an inverse distance metric.
Should be vectorized such that one of \code{a} or \code{b} may be a vector
when the other is a scalar.}
}
\description{
Does not strictly adhere to the guidelines in \link{interpolations}, but can 
be used by \code{\link{getDistanceWeightedInterpolation}} to produce a 
function that does.
}
