% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolations.R
\name{genericSmoothSplineInterpolation}
\alias{genericSmoothSplineInterpolation}
\title{A parameterizable smoothing spline function.}
\usage{
genericSmoothSplineInterpolation(dates.in, y.in, dates.out, ...)
}
\arguments{
\item{dates.in}{A numeric vector desribing the dates for each of the values 
in \code{y.in}. Dates are represented as the number of seconds since 1970.}

\item{y.in}{A vector of values (typically fluxes or concentrations) to 
interpolate among.}

\item{dates.out}{A numeric vector of dates for which the corresponding output
values are to be produced. Dates are represented as the number of seconds 
since 1970.}

\item{...}{any arguments other than \code{x} and \code{y} to be passed to 
\code{stats::\link{smooth.spline}}.}
}
\description{
Does not strictly adhere to the guidelines in \link{interpolations}, but can 
be used by \code{\link{getSmoothSplineInterpolation}} to produce a 
function that does.
}
