% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolations.R
\name{genericTriangularInterpolation}
\alias{genericTriangularInterpolation}
\title{A parameterizable triangular interpolation function.}
\usage{
genericTriangularInterpolation(dates.in, y.in, dates.out, y.mid)
}
\arguments{
\item{dates.in}{A numeric vector desribing the dates for each of the values 
in \code{y.in}. Dates are represented as the number of seconds since 1970.}

\item{y.in}{A vector of values (typically fluxes or concentrations) to 
interpolate among.}

\item{dates.out}{A numeric vector of dates for which the corresponding output
values are to be produced. Dates are represented as the number of seconds 
since 1970.}

\item{y.mid}{The value or values to which midpoints between observations 
should be set. Good choices are: mean(y.in) for stand-alone interpolations;
0 for composite method interpolations among absolute residuals (in log or
linear space); 1 for composite method interpolations among relative
residuals (again, log or linear space).}
}
\description{
Does not strictly adhere to the guidelines in \link{interpolations}, but can 
be used by \code{\link{getTriangularInterpolation}} to produce a 
function that does.
}

