% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{getDateStats}
\alias{getDateStats}
\title{Get various summary statistics on a date vector}
\usage{
getDateStats(dateCol)
}
\arguments{
\item{dateCol}{vector of dates}
}
\value{
data frame of date statistics, including
}
\description{
Get various summary statistics on a date vector
}

