% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolations.R
\name{getDistanceWeightedInterpolation}
\alias{getDistanceWeightedInterpolation}
\title{Generate a distance-weighted interpolation function with the parameters of your choice.}
\usage{
getDistanceWeightedInterpolation(
  inv.dist.fun = function(a, b) {
     1/((a - b)^2)
 }
)
}
\arguments{
\item{inv.dist.fun}{A function to calculate an inverse distance metric.
Should be vectorized such that one of \code{a} or \code{b} may be a vector
when the other is a scalar.}
}
\description{
Produces an interpolation function of the form described in \link{interpolations}.
}
