% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R
\name{getMetadata.loadReg}
\alias{getMetadata.loadReg}
\title{Extracts and imports metadata from an rloadest loadReg model into an object of class 
"metadata"}
\usage{
\method{getMetadata}{loadReg}(load.model)
}
\arguments{
\item{load.model}{a loadReg object}
}
\value{
Object of class "metadata" with slots modified according to the
  metadata contained in load.model
}
\description{
Extracts and imports metadata from an rloadest loadReg model into an object of class 
"metadata"
}
\seealso{
Other getMetadata:
\code{\link[=getMetadata]{getMetadata()}},
\code{\link[=getMetadata.loadModel]{getMetadata.loadModel()}}
}
\concept{getMetadata}
