% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{getPred_baseflow}
\alias{getPred_baseflow}
\title{Create a vector of baseflow rates.}
\usage{
getPred_baseflow(
  data,
  metadata,
  method = c("hysep", "1p digital filter", "2p digital filter"),
  da,
  select,
  alpha,
  BFImax,
  ...
)
}
\arguments{
\item{data}{data.frame with columns as specified by metadata}

\item{metadata}{object of class metadata from which Flow and Date columns 
will be selected.}

\item{method}{character naming the method to be used. May be specified by the
first character or characters, e.g., "h", "1", or "2". See details for a 
description of each algorithm.}

\item{da}{numeric, required for method="hysep". As described in 
?DVstats::hysep, 'the drainage area of the basin in square miles.'}

\item{select}{character, optional argument to method="hysep". As described in
?DVstats::hysep, 'a character string indicating which method to use for the
baseflow in the output dataset. Must be one of "sliding," "local minimum," 
or "fixed." Onle the first letter is required.' See the HYSEP documentation
by Sloto and Crouse at 
https://water.usgs.gov/software/HYSEP/code/doc/hysep.pdf for method 
descriptions.}

\item{alpha}{numeric, required argument for \code{method \%in\% c("1p digital
filter","2p digital filter")}. Adjusts the strength of the filter, with 
numbers closer to 1 leading to more of total flow being classified as 
runoff (non-baseflow). For the 2-parameter digital filter, the default is 
0.98.}

\item{BFImax}{numeric, required argumend for \code{method=="2p digital 
filter"}. The maximum long-term fraction of total streamflow that may be 
classified as baseflow. Recommended values are 0.8 for perennial streams 
with porous aquifers, 0.50 for ephemeral streams with porous aquifers, and 
0.25 for perennial streams with hard rock aquifers. (Defaults from 
https://user.engineering.uiowa.edu/~flood/handouts/ 
HO-L17-Baseflow-Separation.pdf)}

\item{...}{other arguments passed to the specified method (currently only 
'hysep' takes other arguments; see ?DVstats::hysep)}
}
\description{
Internal for now; not yet ready to be a part of the official API
}
\details{
Given a dataset with date and flow columns, separates total flow into 
baseflow and runoff; returns a vector of baseflow rates. Several methods are 
available and can be selected with the "method" argument. All methods assume 
that the time series is evenly spaced, and all work better with 
higher-resolution data (e.g., hourly or daily rather than weekly or monthly).

method=="1p digital filter": Uses the one-paramter digital filter described 
by Nathan, R.J. and T.A. McMahon, 1990. Evaluation of Automated Techniques 
for Baseflow and Recession Analysis. Water Resources Research, 
26(7):1465-1473. This method requires specification of the \code{alpha} 
parameter.

method=="2p digital filter": Uses the two-parameter digital filter described 
by Eckhardt, K., 2005. How to Construct Recursive Digital Filters for 
Baseflow Separation. Hydrological Processes, 19(2):507-515. This method 
requires specification of the \code{alpha} and \code{BFImax} parameters.
}
\references{
Implementation of the 1- and 2-parameter digital filters was 
  aided by the online lecture notes of A. Allen Bradley, Jr. for his Fall 
  2013 Hydrology course (https://user.engineering.uiowa.edu/~flood)
}
\seealso{
Other predictors: 
\code{\link{getPred_isBaseflow}()}
}
\concept{predictors}
\keyword{internal}
