% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R
\name{getPVal}
\alias{getPVal}
\title{helper function to compute the p-value like rloadest does in 
print.loadReg}
\usage{
getPVal(fit)
}
\arguments{
\item{fit}{the loadReg lfit or cfit object}
}
\description{
helper function to compute the p-value like rloadest does in 
print.loadReg
}
\keyword{internal}

