% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{getResiduals}
\alias{getResiduals}
\title{getResiduals return the residuals of the load.model}
\usage{
getResiduals(
  load.model,
  flux.or.conc = c("flux", "conc"),
  abs.or.rel.resids = c("absolute", "relative"),
  use.log = FALSE,
  newdata = NULL,
  raw = FALSE
)
}
\arguments{
\item{load.model}{a loadModel descendant}

\item{flux.or.conc}{The format in which residuals should be calculated}

\item{abs.or.rel.resids}{Should residuals be computed as the difference
("absolute") or the ratio ("relative") of the observed and predicted
values?}

\item{use.log}{logical. Should residuals be computed in log space (TRUE) or
linear space (FALSE)?}

\item{newdata}{New data for prediction and calculation of residuals}

\item{raw}{logical. [not yet implemented] If TRUE, the residuals are
calculated for predictions as they emerge from the inner model. If FALSE,
those outputs are retransformed if appropriate (e.g., from log to linear
space) and converted to the specified format (flux or conc).}
}
\value{
return The residuals between model predictions and the observations
  in newdata
}
\description{
getResiduals return the residuals of the load.model
}
