% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covariance.R
\name{getRhoFirstOrderFun}
\alias{getRhoFirstOrderFun}
\title{Produces a function that uses a first-order autocorrelation model to estimate
the correlation between two dates.}
\usage{
getRhoFirstOrderFun(rho, time.step = as.difftime(1, units = "hours"))
}
\arguments{
\item{rho}{The coefficient of the first-order autocorrelation model, either
asserted by the user or estimated by \code{\link{estimateRho}}.}

\item{time.step}{difftime. The time.step of the time series used to estimate
\code{rho}. It is essential to use the right units; \code{as.difftime(1,
units="hours")} is quite different from \code{as.difftime(1,
units="days")}.}
}
\value{
A function that accepts two date arguments and returns the predicted
  correlation between those two dates. One of the date arguments may be a
  vector as long as the other is scalar. This function takes the form,
  \code{function(date1, date2)} with arguments as defined for
  \code{\link{rhoEqualDates}} and \code{\link{rho1DayBand}}.
}
\description{
Produces a function of the form described in \link{correlations-1D}.

\code{getRhoFirstOrderFun} accepts an assumed or empirically estimated value
of \eqn{\rho} (\code{rho}), the correlation coefficient in the first-order
autocorrelation model:

\deqn{cor(y(t1), y(t2)) = \rho^\tau}

where

\deqn{\tau = y(t2) - y(t1)}

and \eqn{\tau} is in the same units as those used to estimate \eqn{\rho}
(\code{rho}). To ensure that the units are the same, this function also
accepts a \code{time.step} argument indicating the interval between two
successive values in the time series used to estimate \eqn{\rho}
(\code{rho}). The function returned by \code{getRhoFirstOrderFun} takes
responsibility for matching the units of \code{date1} and \code{date2} to
those of \code{rho}.
}

