% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolations.R
\name{getSmoothSplineInterpolation}
\alias{getSmoothSplineInterpolation}
\title{Generate a smoothing spline function with the parameters of your choice.}
\usage{
getSmoothSplineInterpolation(...)
}
\arguments{
\item{...}{any arguments other than \code{x} and \code{y} to be passed to 
\code{stats::\link{smooth.spline}}.}
}
\value{
A function of the form described in \link{interpolations}, i.e., 
  accepting the arguments \code{dates.in}, \code{y.in}, and \code{dates.out} 
  and returning predictions from a smooth spline function for \code{y.out}. 
  That function will use the arguments supplied in \code{...}.
}
\description{
Produces an interpolation function of the form described in \link{interpolations}.
}

