% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolations.R
\name{getTriangularInterpolation}
\alias{getTriangularInterpolation}
\title{Generate a triangular interpolation function with the parameters of your 
choice.}
\usage{
getTriangularInterpolation(y.mid = 0)
}
\arguments{
\item{y.mid}{The value or values to which midpoints between observations 
should be set. Good choices are: mean(y.in) for stand-alone interpolations;
0 for composite method interpolations among absolute residuals (in log or
linear space); 1 for composite method interpolations among relative
residuals (again, log or linear space).}
}
\value{
A function of the form described in \link{interpolations}, i.e., 
  accepting the arguments \code{dates.in}, \code{y.in}, and \code{dates.out} 
  and returning predictions from a triangular interpolation function for 
  \code{y.out}. That function will use the \code{y.mid} argument supplied 
  here.
}
\description{
Produces an interpolation function of the form described in 
\link{interpolations}.
}

