% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{lamprey_discharge}
\alias{lamprey_discharge}
\title{Discharge data for the Lamprey River.}
\description{
Discharge data for the Lamprey River from 10/1/1999 to 11/16/2014.
Discharge in CFS measured every  15 minutes and collected by the US Geological
 Survey, site 1073500, waterdata.usgs.gov. The Lamprey River is an 81-km river 
flowing through southeastern New Hampshire. Its 548-km2 watershed empties into 
the Great Bay estuary.
}

