% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{lamprey_nitrate}
\alias{lamprey_nitrate}
\title{Nitrate data for the Lamprey River.}
\description{
Nitrate data for the Lamprey River from 10/1/1999 to 11/16/2014.
Nitrate is in mg/L and is collected weekly. The Lamprey River is an 81-km river 
flowing through southeastern New Hampshire. Its 548-km2 watershed empties into 
the Great Bay estuary.
Nitrate concentrations have been monitored with weekly and event-based grab 
samples at Packers Falls on the Lamprey since 10 September 1999 and are
measured by SmartChem discrete analyzer (Westco, Brookfield, CT).
}

