% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{loadflex-deprecated-data}
\alias{loadflex-deprecated-data}
\title{Renamed or deprecated datasets for the \pkg{loadflex} package}
\description{
These datasets have been renamed or removed from the package.
}
\section{Datasets}{

  
  \itemize{
  
  \item There aren't any deprecated or defunct datasets yet.
  
  }
}

\seealso{
loadflex-deprecated
}
