% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/unit.conversions.R
\name{loadflex-deprecated-internal}
\alias{loadflex-deprecated-internal}
\alias{generateUnitsData}
\title{Deprecated internal functions in the \pkg{loadflex} package}
\usage{
generateUnitsData()
}
\description{
These functions were formerly private functions in loadflex but are now
deprecated and may be removed in the future. See the Details section for
suggestion on where to look instead.
}
\details{
\code{generateUnitsData} (defunct as of v1.2.0) has been replaced by the
script at data-raw/createUnitsData.R
}
\keyword{internal}
