% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/examples.R
\name{loadflex-deprecated}
\alias{loadflex-deprecated}
\alias{exampleMetadata}
\title{Deprecated functions in the \pkg{loadflex} package}
\usage{
exampleMetadata()
}
\description{
These functions were formerly part of loadflex but are now deprecated and may
be removed in the future. See the Details section for suggestion on where to
look instead.
}
\details{
\code{exampleMetadata} has been replaced by \code{data(eg_metadata)}
}
\seealso{
loadflex-deprecated-data
}
