% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadflex.R
\docType{package}
\name{loadflex}
\alias{loadflex-package}
\alias{loadflex}
\title{Models and Tools for Watershed Flux Estimates}
\description{
loadflex implements several of the most common methods for modeling and predicting watershed solute fluxes and concentrations, including interpolation and regression models, period-weighted averaging, and the composite method. Integrates seamlessly with the USGS's rloadest package and with native R regression models. Offers a uniform interface for any model type, with which you can quickly fit models, generate predictions, and aggregate to monthly or annual values.
}
\section{Models}{


  \itemize{

  \item Linear and period-weighted interpolations: \code{\link{loadInterp}}

  \item Generic regression-based models: \code{\link{loadModel}}

  \item Regression-based models with rloadest (USGS, from LOADEST):
  \code{\link{loadReg2}}

  \item Composite-method-based models: \code{\link{loadComp}}

  }
}

\section{Predictions}{


  \itemize{

  \item Make predictions at the resolution of the predictor data:
  \code{\link{predictSolute}}

  \item Collect predictions into means or totals over longer time periods:
  \code{\link{aggregateSolute}}

  }
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USGS-R/loadflex}
}

}
\author{
\strong{Maintainer}: Alison P. Appling \email{aappling@usgs.gov}

Authors:
\itemize{
  \item Miguel C. Leon
}

Other contributors:
\itemize{
  \item Lindsay R. Carr [contributor]
  \item W. David Watkins [contributor]
}

}
