% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInterp.R
\docType{class}
\name{loadInterp-class}
\alias{loadInterp-class}
\title{A load model class specific to interpolations for flux estimation.}
\description{
loadInterps use a variety of interpolation methods to connect predictions of
y values (usually fluxes, concentrations, or residuals) over time.
}
\section{Slots}{

\describe{
\item{\code{fit}}{the interpolation model to be used.}

\item{\code{MSE}}{numeric. The mean squared error, i.e., the variance of prediction
errors, probably as estimated by leave-one-out cross validation.}
}}

\seealso{
Other load.model.classes: 
\code{\link{loadComp-class}},
\code{\link{loadLm-class}},
\code{\link{loadModel-class}},
\code{\link{loadReg2-class}}
}
\concept{load.model.classes}
