% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\docType{class}
\name{loadModel-class}
\alias{loadModel-class}
\title{A generic load model class.}
\description{
Class and function definitions for very generic load models that you can
create with a single function call (\code{\link{loadModel}}) or extend to
more specific model types (e.g., \code{\link{loadInterp}},
\code{\link{loadReg2}}, \code{\link{loadComp}})
}
\section{Slots}{

\describe{
\item{\code{fit}}{A statistical model, fit to the data and wrapped by the loadModel
class for additional functionality specific to load models.}

\item{\code{pred.format}}{A string indicating the format of predictions (flux or
conc).}

\item{\code{metadata}}{A metadata object describing the load model.}

\item{\code{data}}{The fitting data for the model (fit).}

\item{\code{fitting.function}}{The function used to create or recreate the loadModel,
possibly with new fitting data.}

\item{\code{y.trans.function}}{A function to be applied to the y variable before
fitting the model.}

\item{\code{retrans.function}}{A function to be applied to the y predictions before
returning their values from \code{predictSolute()}.}
}}

\seealso{
Other load.model.classes:
\code{\link{loadComp-class}},
\code{\link{loadInterp-class}},
\code{\link{loadLm-class}},
\code{\link{loadReg2-class}}
}
\concept{load.model.classes}
