% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{lognormal-moments}
\alias{linToLog}
\alias{logToLin}
\alias{lognormal-moments}
\alias{mixedToLog}
\title{Translate means and standard errors/deviations of lognormal distributions
between log and linear space.}
\usage{
logToLin(meanlog, sdlog, mslist)

linToLog(meanlin, sdlin, mslist)

mixedToLog(meanlin, sdlog, meanlog, sdlin)
}
\arguments{
\item{meanlog}{The mean of the distribution on the log scale, with the same
meaning as the \code{meanlog} argument to \code{\link[stats]{rlnorm}}.}

\item{sdlog}{The standard deviation of the distribution on the log scale,
with the same}

\item{mslist}{Optionally, a list containing the meanlog and sdlog as the
first and second elements. If this argument is specified, the first two
arguments should be left missing.}

\item{meanlin}{The mean of the lognormal distribution on the linear scale.}

\item{sdlin}{The standard deviation of the lognormal distribution on the
linear scale.}
}
\description{
\code{logToLin} makes the transformation from the moments of a lognormal
distribution in log space to the moments of the same distribution in linear
space.

\code{linToLog} makes the opposite transformation, from the moments of a
lognormal distribution in linear space to the moments of the same
distribution in log space.

\code{mixedToLog} transforms from a mixed pair - either meanlin and sdlog, or
meanlog and sdlin - and returns meanlog and sdlog.
}
\details{
\subsection{Terminology}{

Let \code{L} be a sample from a lognormal distribution in linear space; for
example, \code{L=rlnorm(1000)} could be a set of predicted flux rates in
kg/day. Then let \code{N} be the corresponding set of predictions in log
space, as they emerge from a model with a logged term on the left-hand side;
for example, the output of \code{predict(lm(log(L_obs) ~ ...))} where
\code{L_obs} is the set of observed flux rates used to calibrate the
\code{lm} model. \code{N} is related to \code{L} by \code{N=log(L) or
L=exp(N)}, but the parameters of their distributions are more complicated.

(It could also be that \code{L} is not a set of predicted flux rates but
simply the uncertainty distribution of a single flux rate prediction. The
same principles apply.)

Given the above definition,

\itemize{

\item \code{meanlin = mean(L) = mean(exp(N)) = exp(mean(N) + sd(N)^2/2)}

\item \code{sdlin = sd(L) = sd(exp(N)) = sqrt((exp(sd(N)^2) - 1) *
exp(2*mean(N) + sd(N)^2))}

\item \code{meanlog = mean(N) = mean(log(L)) = log(mean(L)^2/sqrt(sd(L)^2 +
mean(L)^2))}

\item \code{sdlog = sd(N) = sd(log(L)) = sqrt(log(1 + sd(L)^2/mean(L)^2))}

}

}

Note that meanlin does NOT simply equal exp(mean(N)), and sdlin does NOT
equal exp(sd(N)); this function exists to apply those more complicated
expressions on the right in the above list. Specifically, this function helps
you convert from \code{meanlog=mean(N)}, \code{sdlog=sd(N)} to an appropriate
mean and sd for \code{L} in linear space.
}

