% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match.arg.loadflex.R
\name{match.arg.loadflex}
\alias{match.arg.loadflex}
\title{Require an argument to match the loadflex conventions for that argument name}
\usage{
match.arg.loadflex(arg, choices, several.ok = FALSE)
}
\arguments{
\item{arg}{An argument, passed as a symbol. May be a missing argument to the
calling function.}
}
\value{
The argument after matching to the allowed options for that argument
  name. This may differ from the original value of \code{arg} if \code{arg}
  is a truncated version of one of the character options for the argument
  (identified using \code{pmatch}, as in \code{match.arg}.
}
\description{
This function is modeled on \code{match.arg}, but it is designed to
facilitate consistency in how argument names are used within the
\code{loadflex} package.
}
\details{
The key differences between this function and \code{match.arg} are that this
function (1) decides what the available options are without reference to the
defaults defined in the calling function, and (2) permits atomic but
non-character choices.
}

