% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{meltDates}
\alias{meltDates}
\title{Helper function for plotting: melt dates from wide to long format.}
\usage{
meltDates(plotsols)
}
\arguments{
\item{plotsols}{A data.frame with a DATE column}
}
\description{
We definitely don't want to include this function in the official API, so 
keeping it internal.
}
\keyword{internal}

