% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa_onLoad.R
\name{onBuild}
\alias{onBuild}
\title{On package load, create a private environment in the namespace.}
\usage{
onBuild()
}
\description{
Should not be run except on building the package.
}
\keyword{internal}

