% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotConcCM}
\alias{plotConcCM}
\title{Plot composite method results for concentrations.}
\usage{
plotConcCM(...)
}
\arguments{
\item{...}{arguments passed to plotCM}
}
\description{
Not sure we want to include this function in the official API, so keeping it
internal for now.
}
\keyword{internal}

