% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotConcResidualsCM}
\alias{plotConcResidualsCM}
\title{Plot concentration residuals.}
\usage{
plotConcResidualsCM(...)
}
\description{
Not sure we want to include this function in the official API, so keeping it
internal for now.
}
\keyword{internal}

