% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotLoadResidualsCM}
\alias{plotLoadResidualsCM}
\title{Plot flux residuals.}
\usage{
plotLoadResidualsCM(...)
}
\description{
Not sure we want to include this function in the official API, so keeping it
internal for now.
}
\keyword{internal}
