% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotLoadsCM}
\alias{plotLoadsCM}
\title{Plot composite method results for fluxes.}
\usage{
plotLoadsCM(...)
}
\arguments{
\item{...}{arguments passed to plotCM}
}
\description{
Not sure we want to include this function in the official API, so keeping it
internal for now.
}
\keyword{internal}

