% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotObservationsCM}
\alias{plotObservationsCM}
\title{Plot observations relevant to the composite method.}
\usage{
plotObservationsCM(...)
}
\arguments{
\item{...}{arguments passed to plotCM}
}
\description{
Not sure we want to include this function in the official API, so keeping it
internal for now.
}
\keyword{internal}
