% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotResidualsCM}
\alias{plotResidualsCM}
\title{Plot concentration or flux residuals.}
\usage{
plotResidualsCM(
  type,
  load.model,
  observations,
  resids,
  dateField = "Date",
  verbose = FALSE,
  xObservations = FALSE,
  xFlow = FALSE,
  day.of.year = FALSE
)
}
\arguments{
\item{load.model}{use this to pull names of date column and constituent name}

\item{observations}{these are the observations to plot the residuals for}

\item{dateField}{name of the date field in observations}

\item{verbose}{debugging flag}

\item{day.of.year}{flag if true will plot residuals by the day of year to,}

\item{residuals}{the observation residuals}
}
\description{
Not sure we want to include this function in the official API, so keeping it 
internal for now.
}
\keyword{internal}
