% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadReg.R
\name{resampleCoefficients.loadReg}
\alias{resampleCoefficients.loadReg}
\title{Resample the coefficients from a loadReg model.}
\usage{
resampleCoefficients.loadReg(fit, flux.or.conc)
}
\arguments{
\item{fit}{a loadReg object}

\item{flux.or.conc}{Should the resampling be done for the coefficients
appropriate to flux or those for concentration?}
}
\value{
A new loadReg object with resampled coefficients such that one of
  predConc or predLoad (corresponding to the value of \code{flux.or.conc})
  will make predictions reflecting those new coefficients. No other
  properties of the returned model are guaranteed.
}
\description{
Dives deep into loadReg objects to replace the coefficients for the purpose
of simulating new predictions.
}

