% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R
\name{rlmetricPPCC}
\alias{rlmetricPPCC}
\title{Compute the probability plot correlation coefficient of a loadReg fit}
\usage{
rlmetricPPCC(fit)
}
\arguments{
\item{fit}{the loadReg lfit or cfit object}
}
\description{
Helper function to compute the probability plot correlation coefficient as 
rloadest does in print.loadReg, lines 141-144
}
\keyword{internal}
