% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R
\name{rlmetricPVal}
\alias{rlmetricPVal}
\title{Compute the p-value of a loadReg fit}
\usage{
rlmetricPVal(fit)
}
\arguments{
\item{fit}{the loadReg lfit or cfit object}
}
\description{
Helper function to compute the p-value like rloadest does in print.loadReg
}
\keyword{internal}
