% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R
\name{rlmetricRsq}
\alias{rlmetricRsq}
\title{Compute the [generalized] R squared of a loadReg fit}
\usage{
rlmetricRsq(fit)
}
\arguments{
\item{fit}{the loadReg lfit or cfit object}
}
\description{
Helper function to compute the R squared like rloadest does in print.loadReg
}
\keyword{internal}
