% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{show,metadata-method}
\alias{show,metadata-method}
\title{Display a metadata object}
\usage{
\S4method{show}{metadata}(object)
}
\arguments{
\item{object}{The metadata object to be displayed}
}
\description{
Display a metadata object
}
\examples{
md <- metadata(constituent="NO3", flow="DISCHARGE", 
  dates="DATE", conc.units="mg L^-1", flow.units="cfs", load.units="kg", 
  load.rate.units="kg d^-1", site.name="Lamprey River, NH")
show(md) # or just md at the command prompt
}
