% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R
\name{summarizeModel.loadReg}
\alias{summarizeModel.loadReg}
\title{Extract model summary statistics from an \code{\link[rloadest:loadReg]{rloadest::loadReg()}} model}
\usage{
\method{summarizeModel}{loadReg}(load.model, flux.or.conc = c("flux", "conc"), ...)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel
and always implementing the loadModelInterface.}

\item{flux.or.conc}{character. Which internal model (the flux model or the
concentration model) should be summarized? An \link[rloadest:loadReg]{rloadest::loadReg} model is
actually two different models for (1) flux and (2) concentration, each
fitted to the same data and with the same model structure except for
whether the left-hand side of the model formula is flux or concentration.
Some of the model metrics differ between these two internal models.}

\item{...}{Other arguments passed to model-specific methods}
}
\value{
Returns a 1-row data frame with the following columns:
\itemize{
\item \code{eqn} - the regression equation, possibly in the form \code{const ~ model(x)}
where \code{x} is the \code{Number} of a pre-defined equation in \link[rloadest:Models]{rloadest::Models}
\item \code{RMSE} - the square root of the mean squared error. Errors will be
computed from either fluxes or concentrations, as determined by the value
of \code{pred.format} that was passed to \code{\link[=loadReg2]{loadReg2()}} when this model was
created
\item \code{r.squared} - the r-squared value, generalized for censored data,
describing the amount of observed variation explained by the model
\item \code{p.value} - the p-value for the overall model fit
\item \code{cor.resid} - the serial correlation of the model residuals
\item \code{PPCC} - the probability plot correlation coefficient measuring the
normality of the residuals
\item \code{Intercept}, \code{lnQ}, \code{lnQ2}, \code{DECTIME}, \code{DECTIME2}, \code{sin.DECTIME},
\code{cos.DECTIME}, etc. - the fitted value of the intercept and other terms
included in this model (list differs by model equation)
\item \code{Intercept.SE}, \code{lnQ.SE}, \code{lnQ2.SE}, \code{DECTIME.SE}, \code{DECTIME2.SE},
\code{sin.DECTIME.SE}, \code{cos.DECTIME.SE}, etc. - the standard error of the fitted
estimates of these terms
\item \code{Intercept.p.value}, \code{lnQ.p.value}, \code{lnQ2.p.value}, \code{DECTIME.p.value},
\code{DECTIME2.p.value}, \code{sin.DECTIME.p.value}, \code{cos.DECTIME.p.value} - the
p-values for each of these model terms
}
}
\description{
Produce a 1-row data.frame of model metrics. The relevant metrics for loadReg
models are largely the same as those reported by the \code{rloadest} package,
though reported in this streamlined data.frame format for bulk reporting.
\code{summarizeModel.loadReg} should rarely be accessed directly; instead, call
\code{summarizeModel()} on a \link[rloadest:loadReg]{loadReg} object.
}
