% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizePreds}
\alias{summarizePreds}
\title{Summarize generated predictions 
just take one site for now}
\usage{
summarizePreds(preds, meta, by, modelName)
}
\arguments{
\item{pred}{data frame input data frame of predicted solutes}
}
\value{
data frame containing various statistics
}
\description{
Summarize generated predictions 
just take one site for now
}

