% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeSite}
\alias{summarizeSite}
\title{summarize WQ site information}
\usage{
summarizeSite(site.info, constit.df)
}
\arguments{
\item{site.info}{data frame containing site information, with column for station id}

\item{constit.df}{data frame record of constituent measurements}
}
\value{
data frame of the following information:
}
\description{
summarize WQ site information
}

