% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeSites}
\alias{summarizeSites}
\title{summarize WQ site information}
\usage{
summarizeSites(sites, site.info, nutri.df)
}
\arguments{
\item{sites}{character ID(s) of sites to summarize}

\item{site.info}{data frame containing site information, with column named CODIGO_ESTACAO 
containing station id's}

\item{nutri.df}{data frame record of nutrient measurements}
}
\value{
data frame of the following information:
}
\description{
summarize WQ site information
}

