% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{transformDates}
\alias{transformDates}
\title{Helper function for plotting: transform dates from something to something.}
\usage{
transformDates(plotsols)
}
\arguments{
\item{plotsols}{A data.frame with a DATE column}
}
\value{
same data.frame but with DATE field converted to POSIXct or lt
}
\description{
We definitely don't want to include this function in the official API, so
keeping it internal.
}
\keyword{internal}

