% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{translateFreeformToUnitted}
\alias{translateFreeformToUnitted}
\title{Convert units from a greater variety of forms, including rloadest form, to
unitted form}
\usage{
translateFreeformToUnitted(freeform.units, attach.units = FALSE)
}
\arguments{
\item{freeform.units}{character string or list of character strings
describing one or more sets of units. "Freeform" is an exaggeration but
gets at the idea that these units can take a greater variety of forms than
those accepted by unitted and the units conversion functions in
\code{loadflex}.}

\item{attach.units}{logical. If TRUE, returned value is unitted; otherwise,
it's character.}
}
\value{
a unitbundle or list of unitbundles containing equivalent but simpler
  and more uniform units
}
\description{
Convert units from a greater variety of forms, including rloadest form, to
unitted form
}
\examples{
loadflex:::translateFreeformToUnitted("cfs") # "ft^3 s^-1"
loadflex:::translateFreeformToUnitted("kg/d") # "kg d^-1"
loadflex:::translateFreeformToUnitted("mg L^-1") # "mg L^-1"
}

