% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit.conversions.R
\docType{data}
\name{unit.conversions}
\alias{unit.conversions}
\title{Defines the units that are permissible in load models, along with the
multiplers that will allow common conversions among units}
\format{A data.frame with character columns "numerator", "denominator", and
  numeric column "value", one conversion per row}
\description{
Defines the units that are permissible in load models, along with the
multiplers that will allow common conversions among units
}
\examples{
data(unit.conversions); unit.conversions
}
\keyword{data}
\keyword{units}

