% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\docType{data}
\name{units_loadflex}
\alias{units_loadflex}
\alias{valid.metadata.units}
\alias{freeform.unit.translations}
\alias{unit.conversions}
\title{Units-related datasets for the \pkg{loadflex} package}
\description{
These datasets are provided for managing units within the 
  \pkg{loadflex} package. It is recommended NOT to use these datasets 
  directly within your code, as they may change structure or contents in the 
  future.
}
\section{Datasets}{

  
  \describe{
  
  \item{\code{valid.metadata.units}}{A data.frame with character columns 
  "unit", "standard", and "dimension". Used (1) to determine whether units in
  slots of a `metadata` object have the appropriate dimensionality; e.g., 
  \code{conc.units} must have dimensions of mass or count per volume, and (2)
  to convert among units of the same dimensionality.}
  
  \item{\code{freeform.unit.translations}}{A data.frame with character 
  columns "new" and "old". Used for mapping freeform units ("old") to the 
  specific format required by \pkg{loadflex} ("new").}
  
  \item{\code{unit.conversions}}{A data.frame with character columns 
  "numerator", "denominator", and numeric column "value", one conversion per 
  row. Used for converting among }
  
  }
}

