% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{unitType}
\alias{unitType}
\title{Return the unit.type of a unit string}
\usage{
unitType(unitstr)
}
\arguments{
\item{unitstr}{A string representing units (just one at a time, please)}
}
\description{
Return the unit.type of a unit string
}
\examples{
loadflex:::unitType('kg') # 'load.units'
loadflex:::unitType('kg/d') # NA
loadflex:::unitType(loadflex:::translateFreeformToUnitted('kg/d')) # 'load.rate.units'
loadflex:::unitType('nothing') # NA
}
