% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{valid.metadata.units}
\alias{valid.metadata.units}
\title{Defines the units (or pieces of units) that are permissible in valid load
models}
\format{A data.frame with character columns "unit" and "dimension"}
\description{
Contains a dictionary of valid units and their dimensions
}

