% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{validMetadataUnits}
\alias{validMetadataUnits}
\title{Check whether these units are acceptable (without translation) for inclusion in metadata}
\usage{
validMetadataUnits(
  unitstr,
  unit.type = c("ANY", "flow.units", "conc.units", "load.units", "load.rate.units",
    "basin.area.units")
)
}
\arguments{
\item{unitstr}{A string representing units (just one bundle at a time, please)}

\item{unit.type}{string. accepts "ANY","flow.units","conc.units","load.units", or "load.rate.units"}

\item{type}{A string describing the type of units desired}
}
\value{
logical. TRUE if valid for that unit type, FALSE otherwise
}
\description{
Check whether these units are acceptable (without translation) for inclusion in metadata
}
\examples{
validMetadataUnits("colonies d^-1") # TRUE
validMetadataUnits("m^3 s^-1", unit.type="ANY") # TRUE
validMetadataUnits("nonsensical") # FALSE
validMetadataUnits("g", unit.type="load.units") # TRUE
validMetadataUnits("g", unit.type="flow.units") # FALSE
}
\keyword{units}
