% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadComp.R
\docType{class}
\name{compModel-class}
\alias{compModel-class}
\title{#### compModel class ####
The engine of a loadInterp model.}
\description{
This model class is designed to nest within loadComp objects. The compModel
class is actually a container for two inner models implementing the loadModel
interface: reg.model and resid.model. reg.model will usually be a regression
model, often of type loadReg. resid.model is typically a loadInterp, used to
interpolate among the residuals of the reg.model predictions.
}
\section{Slots}{

\describe{
\item{\code{reg.model}}{The regression model - any model that implements the
loadModelInterface.}

\item{\code{resid.model}}{The interpolation model - any model that implements the
loadModelInterface.}

\item{\code{var.pred.error}}{The mean prediction error, stored as a variance (MSE) in
linear space.}

\item{\code{abs.or.rel.resids}}{character specifying whether interpolation should be
by absolute residuals (obs-pred) or relative residuals (obs/pred).}
}}
\seealso{
Other load.model.fits: \code{\link{interpModel-class}}
}

