% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRET}
\alias{convertToEGRET}
\title{Convert loadflex to EGRET object
 
Convert a loadflex object into an EGRET object for plotting.}
\usage{
convertToEGRET(intdat = NULL, estdat = NULL, preds = NULL, meta = NULL,
  preds.type = "Conc")
}
\arguments{
\item{intdat}{data.frame of interpolation data}

\item{estdat}{data.frame of estimation data}

\item{preds}{data.frame of load predictions}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}

\item{preds.type}{character specifying if the predictions being used are
concentrations ("Conc") or fluxes ("Flux").}
}
\description{
Convert loadflex to EGRET object
 
Convert a loadflex object into an EGRET object for plotting.
}

