% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRET}
\alias{convertToEGRET}
\title{Convert loadflex to EGRET object}
\usage{
convertToEGRET(load.model, estdat = NULL, fitdat = NULL, preds = NULL,
  meta = NULL, preds.type = "Conc")
}
\arguments{
\item{load.model}{a loadflex model object. If this is provided, the only 
other necessary argument is \code{estdat}, and even \code{estdat} is 
optional for some plots (boxConcMonth, plotConcTime, plotConcQ, and 
plotFluxQ). If this is not provided, then at least \code{fitdat} and 
\code{meta} must be provided; \code{estdat} and \code{preds} are also 
necessary for many plots.}

\item{estdat}{data.frame of estimation data}

\item{fitdat}{data.frame of data used to fit a model}

\item{preds}{data.frame of concentration predictions}

\item{meta}{loadflex metadata object; it must include constituent, flow, 
dates, conc.units, site.id, and consti.name}

\item{preds.type}{character specifying if the predictions being used are 
concentrations ("Conc") or fluxes ("Flux"). The only permitted value is
"Conc", and this argument will be leaving soon.}
}
\description{
Convert a loadflex object into an EGRET object for plotting.
}
\examples{
data(lamprey_nitrate)
fitdat <- lamprey_nitrate
data(lamprey_discharge)
estdat <- subset(lamprey_discharge, DATE < as.POSIXct("2012-10-01 00:00:00", tz="EST5EDT")) 
estdat <- estdat[seq(1, nrow(estdat), by=96/4),] # only keep 4 observations per day
meta <- metadata(constituent="NO3", flow="DISCHARGE", dates="DATE", 
  conc.units="mg L^-1", flow.units="cfs", load.units="kg", load.rate.units="kg d^-1", 
  site.name="Lamprey River, NH", site.id='NWIS 01073500', consti.name = "nitrate")
no3_lm <- loadLm(formula=log(NO3) ~ log(DISCHARGE), pred.format="conc", 
  data=fitdat, metadata=meta, retrans=exp)
preds_conc <- predictSolute(no3_lm, "conc", estdat, se.pred=TRUE, date=TRUE)
preds_flux <- predictSolute(no3_lm, "flux", estdat, se.pred=TRUE, date=TRUE)
loadflex:::convertToEGRET(fitdat, estdat, preds_conc, meta)
}

