% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRET}
\alias{convertToEGRET}
\title{Convert loadflex to EGRET object
 
Convert a loadflex object into an EGRET object for plotting.}
\usage{
convertToEGRET(intdat, estdat, preds, meta)
}
\arguments{
\item{intdat}{data.frame of interpolation data}

\item{estdat}{data.frame of estimation data}

\item{preds}{data.frame of load predictions}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: staAbbr)}
}
\description{
Convert loadflex to EGRET object
 
Convert a loadflex object into an EGRET object for plotting.
}

