% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETDaily}
\alias{convertToEGRETDaily}
\title{Convert estimation and load prediction data into the EGRET Daily data.frame}
\usage{
convertToEGRETDaily(estdat, meta, preds, preds.type = "Conc",
  qconvert = 35.314667)
}
\arguments{
\item{estdat}{data.frame of estimation data}

\item{meta}{loadflex metadata object; it must include constituent,
flow, dates, conc.units, site.id, and consti.name}

\item{preds}{data.frame of load predictions}

\item{preds.type}{character specifying if the predictions being used are
concentrations ("Conc") or fluxes ("Flux").}

\item{qconvert}{numeric conversion factor to get flow into cubic meters per second. Default
conversion factor is for cubic feet per second.}
}
\description{
Convert estimation and load prediction data into the EGRET Daily data.frame
}
