% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETDaily}
\alias{convertToEGRETDaily}
\title{Convert estimation and load prediction data into the EGRET Daily data.frame}
\usage{
convertToEGRETDaily(newdata, load.model = NULL, meta = NULL)
}
\arguments{
\item{newdata}{data.frame of data used to generate predictions from an
already-fitted model}

\item{load.model}{a load model (loadReg2, loadComp, loadInterp, loadLm, etc.)
whose data and predictions are to be converted to EGRET format}

\item{meta}{loadflex metadata object; it must include constituent,
flow, dates, conc.units, site.id, and consti.name}
}
\description{
Convert estimation and load prediction data into the EGRET Daily data.frame
}
