% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETInfo}
\alias{convertToEGRETInfo}
\title{Convert a loadflex metadata object into the EGRET INFO dataframe.}
\usage{
convertToEGRETInfo(meta)
}
\arguments{
\item{meta}{loadflex metadata object; it must include site.name, consti.name,
site.id, constituent, and conc.units}
}
\description{
Convert a loadflex metadata object into the EGRET INFO dataframe.
}
