% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETInfo}
\alias{convertToEGRETInfo}
\title{Convert a loadflex metadata object into the EGRET INFO dataframe.}
\usage{
convertToEGRETInfo(meta, preds.type = "Conc")
}
\arguments{
\item{meta}{loadflex metadata object; it must include constituent,
flow, dates, conc.units, site.id, and consti.name}

\item{preds.type}{character specifying if the predictions being used are
concentrations ("Conc") or fluxes ("Flux").}
}
\description{
Convert a loadflex metadata object into the EGRET INFO dataframe.
}

