% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETSample}
\alias{convertToEGRETSample}
\title{Convert the interpolation data.frame into the EGRET Sample dataframe.}
\usage{
convertToEGRETSample(data = NULL, meta = NULL, dailydat = NULL)
}
\arguments{
\item{data}{data.frame of data used to fit a model. only required if
load.model is omitted}

\item{meta}{loadflex metadata object; it must include constituent, flow,
dates, conc.units, site.id, and consti.name. only required if load.model is
omitted}

\item{dailydat}{an EGRET Daily data.frame of flow and prediction values}
}
\description{
Convert the interpolation data.frame into the EGRET Sample dataframe.
}
