% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{convertToEGRETSample}
\alias{convertToEGRETSample}
\title{Convert the interpolation data.frame into the EGRET Sample dataframe.}
\usage{
convertToEGRETSample(intdat, meta, qconvert = 35.314667, dailydat = NULL)
}
\arguments{
\item{intdat}{data.frame of interpolation data}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}

\item{qconvert}{numeric conversion factor to get flow into cubic meters per second. Default
conversion factor is for cubic feet per second.}

\item{dailydat}{an EGRET Daily data.frame of flow values}
}
\description{
Convert the interpolation data.frame into the EGRET Sample dataframe.
}

