% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInterp.R
\name{estimateMSE.loadInterp}
\alias{estimateMSE.loadInterp}
\title{Estimate uncertainty in an interpolation using leave-one-out cross
validation.}
\usage{
\method{estimateMSE}{loadInterp}(
  load.model,
  n.out,
  n.iter = floor(nrow(getFittingData(load.model))/n.out),
  replace,
  ...
)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel
and always implementing the loadModelInterface.}

\item{n.out}{numeric. The number of observations in the fitting data to leave
out in each iteration.}

\item{n.iter}{numeric. The number of iterations to perform.}

\item{replace}{logical. In each iteration, should the n.out observations that
are left out be sampled with replacement of any previous sets of n.out
observations (TRUE) or without replacement (FALSE)?}

\item{...}{Other arguments passed to inheriting methods for estimateMSE}
}
\description{
Calculates and returns the mean squared errors (MSEs) in the units and
transformation space (e.g., log space) of the left-hand side of the model fit
equation.
}
\details{
This method is leave-one-out cross validation (LOOCV) when n.out==1,
n.iter==nrow(getFittingData(load.model)), and with.replacement=FALSE. This
method is k-fold cross validation when
n.out*n.iter==nrow(getFittingData(load.model)) and with.replacement==FALSE.
}
\seealso{
Other estimateMSE: 
\code{\link{estimateMSE}()},
\code{\link{estimateMSE.loadComp}()}
}
\concept{estimateMSE}
