% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimateRho}
\alias{estimateRho}
\title{Estimate the autocorrelation of a mid- to high-resolution time series}
\usage{
estimateRho(load.model, flux.or.conc = c("flux", "conc"),
  abs.or.rel.resids = c("absolute", "relative"), use.log = FALSE,
  newdata = NULL, plot.acf = TRUE, timestep.tol = .Machine$double.eps^0.5,
  irregular.timesteps.ok = FALSE)
}
\arguments{
\item{load.model}{a loadModel descendant}

\item{flux.or.conc}{The format in which residuals should be calculated}

\item{abs.or.rel.resids}{Should residuals be computed as the difference or
the ratio of the observed and predicted values?}

\item{newdata}{prediction values}
}
\value{
Return the rho function and the fitted model.
}
\description{
Uses arima() to estimate the first-order autocorrelation constant, rho, for
the residuals calculated from a load model and new data.
}
\details{
For the purpose of estimating prediction error for a composite method
application, this function should be called with a dataset of comparable (or
identical) resolution to that of the interpolation data.

For the purpose of estimating the covariance of prediction errors for
aggregation, this function should not be called unless newdata are at
approximately the same (or higher) temporal resolution as that of the data to
be used for load estimation. This will almost always involve sensor data.

If you have no data with sufficient resolution to reasonably call this
function, take your pick from the established assumptions implemented in
functions such as \code{\link{rhoEqualDates}} or \code{\link{rho1DayBand}},
or write your own assumptions.
}
\seealso{
Other diagnostics: \code{\link{getCorrectionFraction}},
  \code{\link{getCorrectionFraction.loadComp}};
  \code{\link{residDurbinWatson}}
}

