% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{flowconcToFluxConversion}
\alias{flowconcToFluxConversion}
\title{Provide the conversion factor which, when multiplied by flow * conc, gives
the flux in the desired units}
\usage{
flowconcToFluxConversion(flow.units, conc.units, load.rate.units,
  attach.units = FALSE)
}
\arguments{
\item{flow.units}{character. The units of flow.}

\item{conc.units}{character. The units of concentration.}

\item{load.rate.units}{character. The units of flux.}

\item{attach.units}{logical. If TRUE, the conversion factor is returned with
units attached.}
}
\value{
numeric, or unitted numeric if unitted=TRUE. The conversion factor.
}
\description{
By dividing rather than multiplying by this factor, the output of this
function may also be used to convert from flux units to the units of the
product of flow and concentration.
}
\examples{
flowconcToFluxConversion("L/d", "g/L", "g/d") # 1
flowconcToFluxConversion("cfs", "g/L", "kg/d") # 2446.589
library(unitted); u(10, "ft^3 s^-1") * u(2, "mg L^-1") * flowconcToFluxConversion("cfs", "mg/L", "kg/d", attach.units=TRUE) # u(48.9 ,"kg d^-1")
}

