% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{flowCorrectionEGRET}
\alias{flowCorrectionEGRET}
\title{Use EGRET functions to correct flow values}
\usage{
flowCorrectionEGRET(flowdat, flow.colname, qconvert = 35.314667)
}
\arguments{
\item{flowdat}{data frame with discharge values to convert}

\item{flow.colname}{character string giving the column name that corresponds to flow}

\item{qconvert}{numeric conversion factor to get flow into cubic meters per second. Default
conversion factor is for cubic feet per second.}
}
\description{
Uses the EGRET function populateDaily to convert values 
into cubic meters per second.
}

