% Generated by roxygen2 (4.0.2): do not edit by hand
\name{formatPreds}
\alias{formatPreds}
\title{formatPreds raw to final predictions}
\usage{
formatPreds(preds, from.format = c("flux", "conc*flow", "flux/flow", "conc"),
  to.format = c("flux", "conc"), newdata, metadata, attach.units = FALSE)
}
\arguments{
\item{preds}{raw prediction values}

\item{from.format}{format of the raw predictions}

\item{to.format}{character. Determines if the returned value is a load or a concentration.}

\item{newdata}{}

\item{metadata}{- will be used to determine the units}

\item{attach}{units - attach the units to the returned value}
}
\description{
Convert raw predictions to final predictions, possibly including a switch
between flux and conc. If there is a switch, the units will be converted
according to the metadata.
}

