% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{formatPreds}
\alias{formatPreds}
\title{formatPreds raw to final predictions}
\usage{
formatPreds(preds, from.format = c("flux", "conc*flow", "flux/flow", "conc"),
  to.format = c("flux", "conc"), newdata, metadata, attach.units = FALSE)
}
\arguments{
\item{preds}{raw prediction values}

\item{from.format}{format of the raw predictions}

\item{to.format}{character. Determines if the returned value is a load or a
concentration.}

\item{newdata}{a data.frame with nrow() == length(preds) and with any columns
that will be needed to perform the requested conversion: for example,
from="conc" and to="flux" implies that discharge will be available in
newdata}

\item{metadata}{- will be used to determine the units}

\item{attach}{units - attach the units to the returned value}
}
\value{
converted predictions (in the format/units specified by to.format and
  metadata)
}
\description{
Convert raw predictions to final predictions, possibly including a switch
between flux and conc. If there is a switch, the units will be converted
according to the metadata.
}
\examples{
obs <- transform(data.frame(MyConc=1:10, MyFlow=rep(10,10)), MyFlux=MyConc*MyFlow*rloadest::loadConvFactor("cms", "mg/l", "kg") )
md <- updateMetadata(exampleMetadata(), constituent="MyConc", flow="MyFlow", load.rate="MyFlux", dates="none",
  flow.units="cms", conc.units="mg/l", load.units="kg", load.rate.units="kg/d", custom=NULL)

formatPreds(preds=obs$MyConc, from.format="conc", to.format="flux", newdata=obs, metadata=md) # == obs$MyFlux
formatPreds(preds=obs$MyConc*obs$MyFlow, from.format="conc*flow", to.format="flux", newdata=obs, metadata=md) # == obs$MyFlux
formatPreds(preds=obs$MyFlux, from.format="flux", to.format="conc", newdata=obs, metadata=md) # == obs$MyConc
formatPreds(preds=obs$MyFlux, from.format="flux", to.format="conc", newdata=obs, metadata=md, attach.units=TRUE) # == u(obs$MyConc, "mg L^-1")
}
\keyword{units}

