% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{formatPreds}
\alias{formatPreds}
\title{formatPreds raw to final predictions}
\usage{
formatPreds(preds, from.format = c("flux", "conc*flow", "flux/flow", "conc"),
  to.format = c("flux", "conc"), newdata, metadata, attach.units = FALSE)
}
\arguments{
\item{preds}{raw prediction values}

\item{from.format}{character in 
\code{c("flux","conc*flow","flux/flow","conc")}. Format of the raw 
predictions.}

\item{to.format}{character indicating whether the returned value should be a 
flux or a concentration.}

\item{newdata}{a data.frame with nrow() == length(preds) and containing any 
columns (named as in \code{metadata}) that will be needed to perform the
requested conversion. For example, from="conc" and to="flux" implies that
a discharge column will be available in \code{newdata}.}

\item{metadata}{An object of class \code{\link{metadata}} used to determine 
the units of inputs and desired output}

\item{attach.units}{logical. Attach the units to the returned value?}
}
\value{
converted predictions (in the format/units specified by to.format and
  metadata)
}
\description{
Convert raw predictions to final predictions, possibly including a switch 
between flux and conc. If there is a switch, the units will be converted 
according to the metadata.
}
\examples{
obs <- transform(data.frame(MyConc=1:10, MyFlow=rep(10,10)), 
MyFlux=MyConc*MyFlow*rloadest::loadConvFactor("cms", "mg/l", "kg") )
md <- updateMetadata(exampleMetadata(), constituent="MyConc", flow="MyFlow", 
load.rate="MyFlux", dates="none", flow.units="cms", conc.units="mg/l", load.units="kg", 
load.rate.units="kg/d", custom=NULL)
  
formatPreds(preds=obs$MyConc, from.format="conc", to.format="flux", newdata=obs, 
metadata=md) # == obs$MyFlux
formatPreds(preds=obs$MyConc*obs$MyFlow, from.format="conc*flow", to.format="flux", newdata=obs, 
metadata=md) # == obs$MyFlux
formatPreds(preds=obs$MyFlux, from.format="flux", to.format="conc", newdata=obs, 
metadata=md) # == obs$MyConc
formatPreds(preds=obs$MyFlux, from.format="flux", to.format="conc", newdata=obs, metadata=md, 
attach.units=TRUE) # == u(obs$MyConc, "mg L^-1")
}
\keyword{units}
