% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covariance.R
\name{getCormatFirstOrder}
\alias{getCormatFirstOrder}
\title{get rho matrix first order}
\usage{
getCormatFirstOrder(rho, time.step = as.difftime(1, units = "hours"),
  max.tao = as.difftime(1, units = "days"))
}
\arguments{
\item{rho}{the covariance  asdefined as difference between the times divided by the time step}

\item{time.step}{default is 1 hour}

\item{max.tao}{don't consider covariance for values further apart then this.}
}
\value{
covariance defined as difference between the times divided by the time step don't calculate covariance for
        values further away then max.tao.
}
\description{
get rho matrix first order
}

