% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadComp.R
\name{getCorrectionFraction}
\alias{getCorrectionFraction}
\alias{getCorrectionFraction.loadComp}
\title{The fraction of prediction that is due to a correction.}
\usage{
getCorrectionFraction(
  load.model,
  flux.or.conc = c("flux", "conc"),
  newdata,
  ...
)

\method{getCorrectionFraction}{loadComp}(
  load.model,
  flux.or.conc = c("flux", "conc"),
  newdata,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{load.model}{The load model from which to pull the correction fraction.}

\item{flux.or.conc}{character. Should the metric be computed with respect to
predictions of flux rate or concentration?}

\item{newdata}{The data for which to compute the correction fraction, usually
containing daily or instantaneous discharge.}

\item{...}{Other arguments passed to class-specific implementations of
getCorrectionFraction.}

\item{na.rm}{logical. Should predictions with NA values be excluded?}
}
\value{
A value between 0 and 1 indicating the fraction of total load that is
  attributed to a correction.
}
\description{
Computes the fraction of total concentration or flux prediction that is
attributable to a correction such as the residuals correction of composite
method models.

\code{\link{loadComp}} models: The correction fraction for a composite method
(\code{loadComp}) model is defined here as:

\deqn{CF = (\sum |R| \Delta t) / (\sum L \Delta t)}

where \eqn{\sum} indicates a sum over all predictions, \eqn{L} is the vector
of composite predictions of flux or concentration, \eqn{R} is the vector of
interpolated residuals in the same format (flux or concentration, non-log
space) as \eqn{L}, and \eqn{\Delta t} is the vector of time periods
represented by the predictions.
}
\seealso{
Other diagnostics:
\code{\link[=estimateRho]{estimateRho()}},
\code{\link[=residDurbinWatson]{residDurbinWatson()}}
}
\concept{diagnostics}
