% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{getFittedModel.loadModel}
\alias{getFittedModel.loadModel}
\title{Retrieve the fitted model, if appropriate, from a loadModel load model}
\usage{
\method{getFittedModel}{loadModel}(load.model)
}
\arguments{
\item{load.model}{The load model for which to return the inner fitted model.}
}
\description{
Retrieve the fitted model, if appropriate, from a loadModel load model
}
\seealso{
Other getFittedModel: 
\code{\link{getFittedModel}()}
}
\concept{getFittedModel}
