% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelInterface.R
\name{getFittedModel}
\alias{getFittedModel}
\title{Retrieve the fitted model, if appropriate, from a loadModel load model}
\usage{
getFittedModel(load.model)
}
\arguments{
\item{load.model}{The load model for which to return the inner fitted model.}
}
\value{
Object of class "function" which
}
\description{
A function in the loadModelInterface. Takes a load.model and returns a
function to fit a new load.model that is identical in every respect except
its training data and resulting model coefficients or other paramters. The
returned function should accept exactly one argument, the training data, and
should return an object of the same class as load.model.
}
\seealso{
Other loadModelInterface: 
\code{\link{estimateMSE}()},
\code{\link{getFittingData}()},
\code{\link{getFittingFunction}()},
\code{\link{getMetadata}()},
\code{\link{predictSolute}()},
\code{\link{simulateSolute}()},
\code{\link{summarizeModel}()},
\code{\link{validLoadModelInterface}()}

Other getFittedModel: 
\code{\link{getFittedModel.loadModel}()}
}
\concept{getFittedModel}
\concept{loadModelInterface}
