% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{getFittingData.loadModel}
\alias{getFittingData.loadModel}
\title{Retrieve the data used to fit the model}
\usage{
\method{getFittingData}{loadModel}(load.model)
}
\arguments{
\item{load.model}{A load model, implementing the loadModelInterface, for
which to return the fitting data}
}
\description{
Retrieve the data used to fit the model
}
\seealso{
Other getFittingData: 
\code{\link{getFittingData}()}
}
\concept{getFittingData}
