% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFittingData.loadModel}
\alias{getFittingData.loadModel}
\title{Retrieve the data used to fit the model}
\usage{
\method{getFittingData}{loadModel}(load.model)
}
\arguments{
\item{load.model}{A load model, implementing the loadModelInterface, for
which to return the fitting data}
}
\description{
Retrieve the data used to fit the model
}
\seealso{
Other getFittingData: \code{\link{getFittingData}}
}

