% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelInterface.R
\name{getFittingData}
\alias{getFittingData}
\title{Extract the data originally used to fit a load model.}
\usage{
getFittingData(load.model)
}
\arguments{
\item{load.model}{A load model, implementing the loadModelInterface, for
which to return the fitting data}
}
\value{
data.frame containing the original fitting data
}
\description{
A function in the loadModelInterface.
}
\seealso{
Other loadModelInterface:
\code{\link[=estimateMSE]{estimateMSE()}},
\code{\link[=getFittedModel]{getFittedModel()}},
\code{\link[=getFittingFunction]{getFittingFunction()}},
\code{\link[=getMetadata]{getMetadata()}},
\code{\link[=predictSolute]{predictSolute()}},
\code{\link[=simulateSolute]{simulateSolute()}},
\code{\link[=summarizeModel]{summarizeModel()}},
\code{\link[=validLoadModelInterface]{validLoadModelInterface()}}

Other getFittingData:
\code{\link[=getFittingData.loadModel]{getFittingData.loadModel()}}
}
\concept{getFittingData}
\concept{loadModelInterface}
