% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadModel.R
\name{getFittingFunction.loadModel}
\alias{getFittingFunction.loadModel}
\title{Retrieve a fitting function from a loadModel load model}
\usage{
\method{getFittingFunction}{loadModel}(load.model)
}
\arguments{
\item{load.model}{The model for which to return a new fitting function.}
}
\description{
Retrieve a fitting function from a loadModel load model
}
\seealso{
Other getFittingFunction: \code{\link{getFittingFunction}}
}

