% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelInterface.R
\name{getFittingFunction}
\alias{getFittingFunction}
\title{Get a function that can be used to refit the load.model with new data.}
\usage{
getFittingFunction(load.model)
}
\arguments{
\item{load.model}{The model for which to return a new fitting function.}
}
\value{
Object of class "function" which
}
\description{
A function in the loadModelInterface. Takes a load.model and returns a
function to fit a new load.model that is identical in every respect except
its training data and resulting model coefficients or other paramters. The
returned function should accept exactly one argument, the training data, and
should return an object of the same class as load.model.
}
\seealso{
Other loadModelInterface: 
\code{\link{estimateMSE}()},
\code{\link{getFittedModel}()},
\code{\link{getFittingData}()},
\code{\link{getMetadata}()},
\code{\link{predictSolute}()},
\code{\link{simulateSolute}()},
\code{\link{summarizeModel}()},
\code{\link{validLoadModelInterface}()}

Other getFittingFunction: 
\code{\link{getFittingFunction.loadModel}()}
}
\concept{getFittingFunction}
\concept{loadModelInterface}
