% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{getMetadata.loadModel}
\alias{getMetadata.loadModel}
\title{Retrieve metadata from a loadModel load model}
\usage{
\method{getMetadata}{loadModel}(load.model)
}
\arguments{
\item{load.model}{A load model, implementing the loadModelInterface, for
which to return the metadata}
}
\description{
Retrieve metadata from a loadModel load model
}
\seealso{
Other getMetadata:
\code{\link[=getMetadata]{getMetadata()}},
\code{\link[=getMetadata.loadReg]{getMetadata.loadReg()}}
}
\concept{getMetadata}
