% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadReg.R
\name{getMetadata.loadReg}
\alias{getMetadata.loadReg}
\title{Extracts and imports metadata from an rloadest loadReg model into an object of class
"metadata"}
\usage{
\method{getMetadata}{loadReg}(fit)
}
\arguments{
\item{load.model}{object of class "loadReg", typically produced using
rloadest::loadReg(), from which metadata is to be imported}
}
\value{
Object of class "metadata" with slots modified according to the
  metadata contained in load.model
}
\description{
Extracts and imports metadata from an rloadest loadReg model into an object of class
"metadata"
}
\seealso{
Other getMetadata: \code{\link{getMetadata.loadModel}};
  \code{\link{getMetadata}}
}

