% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelInterface.R
\name{getMetadata}
\alias{getMetadata}
\title{Extract metadata from a load model.}
\usage{
getMetadata(load.model)
}
\arguments{
\item{load.model}{A load model, implementing the loadModelInterface, for
which to return the metadata}
}
\value{
Object of class "metadata" with slots reflecting the metadata for
  load.model
}
\description{
A function in the loadModelInterface. Returns a load model's metadata,
encapsulated as a \code{metadata} object.
}
\seealso{
Other loadModelInterface:
\code{\link[=estimateMSE]{estimateMSE()}},
\code{\link[=getFittedModel]{getFittedModel()}},
\code{\link[=getFittingData]{getFittingData()}},
\code{\link[=getFittingFunction]{getFittingFunction()}},
\code{\link[=predictSolute]{predictSolute()}},
\code{\link[=simulateSolute]{simulateSolute()}},
\code{\link[=summarizeModel]{summarizeModel()}},
\code{\link[=validLoadModelInterface]{validLoadModelInterface()}}

Other getMetadata:
\code{\link[=getMetadata.loadModel]{getMetadata.loadModel()}},
\code{\link[=getMetadata.loadReg]{getMetadata.loadReg()}}
}
\concept{getMetadata}
\concept{loadModelInterface}
