% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\name{getPred_isBaseflow}
\alias{getPred_isBaseflow}
\title{Create a logical vector indicating whether flow at each time point is 
dominantly baseflow.}
\usage{
getPred_isBaseflow(data, metadata, baseflow, threshold = 0.8)
}
\arguments{
\item{data}{a data.frame of data including a column for flow (with that 
column name specified by metadata).}

\item{metadata}{a metadata object specifying, at minimum, the column name in 
\code{data} where total flow is stored.}

\item{baseflow}{a vector of baseflow values of the same length as 
\code{nrow(data)}.}

\item{threshold}{numeric. The minimum fraction of flow that must be baseflow 
for a time point to be classified as baseflow (TRUE).}
}
\description{
Internal for now; not yet ready to be a part of the official API
}
\seealso{
Other predictors: \code{\link{getPred_baseflow}}
}
\keyword{internal}

