% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{getPVal}
\alias{getPVal}
\title{helper function to compute the p-value like rloadest does in 
print.loadReg}
\usage{
getPVal(model)
}
\arguments{
\item{model}{the loadReg model object}
}
\description{
helper function to compute the p-value like rloadest does in 
print.loadReg
}

