% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInterp.R
\docType{class}
\name{interpModel-class}
\alias{interpModel-class}
\title{loadInterp is a class of load models that hold interpolation functions.
The engine of a loadInterp model.}
\description{
This is a model class to nest within loadInterp. This class, interpModel, is 
lightweight relative to loadInterp: its focus is on the interpolation rather 
than on units conversions, fitting, or other things that loadInterp does 
better. This is also the class that is produced by a call to the fitting 
function stored in loadInterp.
}
\section{Slots}{

\describe{
\item{\code{dates.class}}{Class of the dates.in.}

\item{\code{dates.in}}{Dates of the y.in data.}

\item{\code{y.in}}{Data (usually fluxes, concentrations, or residuals) to 
interpolate.}

\item{\code{interp.function}}{Function that accepts arguments \code{dates.in}, 
\code{y.in}, and \code{dates.out}, then returns predictions \code{y.out}}
}}

\seealso{
Other load.model.fits: \code{\link{compModel-class}}
}
\concept{load.model.fits}
