% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{isTimestepRegular}
\alias{isTimestepRegular}
\title{Check a time series for evenly spaced dates.}
\usage{
isTimestepRegular(dates, hist = TRUE, tol = .Machine$double.eps^0.5,
  handler = stop)
}
\arguments{
\item{dates}{A time series of dates in any format accepted by \code{diff}.}

\item{hist}{logical. If the time series is found to be irregular, should this
function print a histogram of the observed time steps?}

\item{tol}{time step tolerance, the accepatable amount of difference between time steps to still consider them regular.}

\item{handler}{A function, e.g., \code{stop} or \code{warning}, to be run if
the time series is irregular.}
}
\value{
TRUE if the time steps in \code{dates} are identical to within the 
  tolerance set by {tol}, FALSE otherwise.
}
\description{
Even spacing of dates is required for many tests and assessments of 
autocorrelation. This function checks for even spacing to within a specified 
tolerance. If spacing is uneven, a histogram is immediately produced (if 
\code{hist=TRUE}), and a handling function is run (e.g., to produce an error 
or warning, as specified by \code{handler}).
}
