% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadComp.R
\docType{class}
\name{loadComp-class}
\alias{loadComp-class}
\title{A load model class implementing the composite method for flux estimation.}
\description{
A load model class implementing the composite method for flux estimation.
}
\seealso{
Other load.model.classes: \code{\link{loadInterp-class}},
  \code{\link{loadLm-class}},
  \code{\link{loadModel-class}},
  \code{\link{loadReg2-class}}
}
