% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadflex.R
\docType{_PACKAGE}
\name{loadflex}
\alias{loadflex}
\title{Models and Tools for Watershed Flux Estimates}
\description{
Models and Tools for Watershed Flux Estimates
}
\section{Models}{


  \itemize{

  \item Linear and period-weighted interpolations: \code{\link{loadInterp}}

  \item Generic regression-based models: \code{\link{loadModel}}

  \item Regression-based models with rloadest (USGS, from LOADEST):
  \code{\link{loadReg2}}

  \item Composite-method-based models: \code{\link{loadComp}}

  }
}

\section{Predictions}{


  \itemize{

  \item Make predictions at the resolution of the predictor data:
  \code{\link{predictSolute}}

  \item Collect predictions into means or totals over longer time periods:
  \code{\link{aggregateSolute}}

  }
}

