% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLm.R
\docType{class}
\name{loadLm-class}
\alias{loadLm-class}
\title{A load model class specific to simple linear models (\code{\link{lm}}s) for
flux estimation.}
\description{
loadLms can take any lm formula.
}
\section{Slots}{

\describe{
\item{\code{fit}}{the interpolation model to be used.}

\item{\code{ylog}}{logical. If TRUE, this constitutes affirmation that the values
passed to the left-hand side of the model formula will be in log space. If
missing, the value of \code{ylog} will be inferred from the values of
\code{formula} and \code{y.trans.function}, but be warned that this
inference is fallible.}
}}

\seealso{
Other load.model.classes: 
\code{\link{loadComp-class}},
\code{\link{loadInterp-class}},
\code{\link{loadModel-class}},
\code{\link{loadReg2-class}}
}
\concept{load.model.classes}
