% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadLm.R
\docType{class}
\name{loadLm-class}
\alias{loadLm-class}
\title{A load model class specific to simple linear models (\code{\link{lm}}s) for
flux estimation.}
\description{
loadLms can take any lm formula.
}
\section{Slots}{

\describe{
\item{\code{fit}}{the interpolation model to be used.}
}}
\seealso{
Other load.model.classes: \code{\link{loadComp-class}};
  \code{\link{loadInterp-class}};
  \code{\link{loadModel-class}};
  \code{\link{loadReg2-class}}
}

