% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadLm.R
\name{loadLm}
\alias{loadLm}
\title{Create a fitted loadLm object.}
\usage{
loadLm(formula, pred.format = c("flux", "conc"), data, metadata,
  fitting_function = NULL, y.trans.function = NULL,
  retrans.function = exp, store = c("data", "fitting.function"), ylog)
}
\arguments{
\item{formula}{A formula specifying the linear model to fit.}

\item{pred.format}{In what format (flux or conc) does the lm object make
predictions when predict.lm() is called?}

\item{data}{data.frame. The data to be interpolated}

\item{metadata}{metadata, used to access the appropriate columns of data. At
a minimum, \code{metadata} should correctly specify the date column.}

\item{retrans.function}{}

\item{store}{One or more character strings specifying which information to
write within the model. Options are 'data': the original fitting data;
'fitting.function': a fitting function that can produce a new loadComp
object from new data}

\item{ylog}{logical. If TRUE, this constitutes affirmation that the values
passed to the left-hand side of the model formula will be in log space. If
missing, the value of \code{ylog} will be inferred from the values of
\code{formula} and \code{y.trans.function}, but be warned that this
inference is fallible.}
}
\value{
A fitted loadLm model.
}
\description{
Generates a new model of class loadInterp (\code{\link{loadLm-class}}) which
can connect observations of concentration or flux using a linear regression
model.
}
\seealso{
Other load.model.inits: \code{\link{loadComp}};
  \code{\link{loadInterp}}; \code{\link{loadModel}};
  \code{\link{loadReg2}}
}

