% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModel.R
\name{loadModel}
\alias{loadModel}
\title{Create a fitted loadModel object.}
\usage{
loadModel(
  inner.fit.function,
  pred.format,
  data,
  metadata,
  y.trans.function = NULL,
  retrans.function = exp,
  store = c("data", "fitting.function")
)
}
\arguments{
\item{inner.fit.function}{function that accepts one argument, a training data
frame, and returns a [re]fitted statistical model, such as an \code{lm},
relating flux or concentration to predictors in the training data.}

\item{pred.format}{character string that describes the output of
retrans.function(predict(fit)) [NOT of predictSolute.loadModel(), which can
do either!]}

\item{data}{data.frame containing the initial training observations}

\item{metadata}{An object of class \code{\link{metadata}}}

\item{y.trans.function}{function that accepts a vector of observed response
values (e.g., concentrations or flux rates) and transforms them into the
values on the left-hand side of the calibration formula. Because load
models are frequently fit to log(y) ~ ..., the default is
y.trans.function=log.}

\item{retrans.function}{function that accepts a vector of predictions
straight from the fitted model and retransforms them, if appropriate, to
predictions in linear space and with units of concentration*flow. Because
load models are frequently fit to log(y) ~ ..., the default is
retrans.function=exp. After retrans.function has been applied to the
predictions, the retransformed predictions will automatically undergo
further units conversion from conc*flow to load.rate, according to the
metadata elements conc.units, flow.units, and load.rate.units.}

\item{store}{One or more character strings specifying which information to
write within the model. Options are 'data': the original fitting data;
'fitting.function': a fitting function that can produce a new loadComp
object from new data}
}
\value{
A fitted loadModel.
}
\description{
Generates a new model of class loadModel (\code{\link{loadModel-class}}).
}
\seealso{
Other load.model.inits:
\code{\link[=loadComp]{loadComp()}},
\code{\link[=loadInterp]{loadInterp()}},
\code{\link[=loadLm]{loadLm()}},
\code{\link[=loadReg2]{loadReg2()}}
}
\concept{load.model.inits}
