% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{loadReg2-class}
\alias{loadReg2-class}
\title{A load model class specific to loadReg objects produced by the USGS rloadest
package.}
\description{
This class is closely related to loadModel except that the model slot is more
strictly defined, and the associated methods are a little different.
}
\seealso{
Other load.model.classes: \code{\link{loadComp-class}};
  \code{\link{loadInterp-class}};
  \code{\link{loadLm-class}}; \code{\link{loadModel-class}}
}

