% Generated by roxygen2 (4.0.2): do not edit by hand
\name{metadata-class}
\alias{metadata-class}
\title{Store metadata relevant to a load model.}
\description{
\code{metadata} objects store metadata for a load model of any type (e.g.,
\code{\link{loadModel}}, \code{\link{loadInterp}}, \code{\link{loadReg2}}, or
\code{\link{loadComp}}). Metadata can be created and modified by calls to
\code{\link{metadata}}, \code{\link{updateMetadata}}, or
\code{\link{exampleMetadata}}. Values and relevant information can be
accessed by \code{\link{getCol}}, \code{\link{getUnits}}, or
\code{\link{getInfo}}.
}
\section{Slots}{

\describe{
\item{\code{constitutent}}{character. The name of the data.frame column describing
the concentration of the solute or material to be modeled.}

\item{\code{flow}}{character. The name of the data.frame column describing flow
(discharge)}

\item{\code{load.rate}}{character. The name of the data.frame column, if it exists,
describing instantaneous load}

\item{\code{dates}}{character. The name of the data.frame column describing date
and/or time}

\item{\code{flow.units}}{character. The units in which flow is both input and
exported.}

\item{\code{conc.units}}{character. The units in which concentration is both input
and exported.}

\item{\code{load.units}}{character. The units in which load (flux) is both input and
exported.}

\item{\code{load.rate.units}}{character. The units in which load (flux) is reported
by predict.loadModel().}

\item{\code{station}}{character. A description of the sampling station or site.}

\item{\code{custom}}{ANY. Empty by default, but may be modified to store any
additional data the user wants to track.}
}}

