% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{class}
\name{metadata-class}
\alias{metadata-class}
\title{Store metadata relevant to a load model.}
\description{
\code{metadata} objects store metadata for a load model of any type (e.g., 
\code{\link{loadModel}}, \code{\link{loadInterp}}, \code{\link{loadReg2}}, or
\code{\link{loadComp}}). Metadata can be created and modified by calls to 
\code{\link{metadata}}, \code{\link{updateMetadata}}, or 
\code{\link{exampleMetadata}}. Values and relevant information can be 
accessed by \code{\link{getCol}}, \code{\link{getUnits}}, or 
\code{\link{getInfo}}.
}
\section{Slots}{

\describe{
\item{\code{constituent}}{character. The name of the data.frame column describing the
concentration of the solute or material to be modeled.}

\item{\code{consti.name}}{character. Consituent long name, for use in plots and 
reports}

\item{\code{flow}}{character. The name of the data.frame column describing flow 
(discharge)}

\item{\code{load.rate}}{character. The name of the data.frame column, if it exists, 
describing instantaneous load}

\item{\code{dates}}{character. The name of the data.frame column describing date 
and/or time}

\item{\code{flow.units}}{character. The units in which flow is both input and 
exported.}

\item{\code{conc.units}}{character. The units in which concentration is both input 
and exported.}

\item{\code{load.units}}{character. The units in which load (flux) is both input and 
exported.}

\item{\code{load.rate.units}}{character. The units in which load (flux) is reported 
by predict.loadModel().}

\item{\code{station}}{deprecated. character. A description of the sampling station or
site. It's now recommended to use \code{site.name} and/or \code{site.id} 
instead.}

\item{\code{site.name}}{character. A description of the sampling station or site. 
Formerly named \code{station}.}

\item{\code{site.id}}{character Station ID for the site where the concentration of 
the constituent was measured.}

\item{\code{lat}}{numeric Station latitude where concentration (and possibly also 
discharge) was measured.}

\item{\code{lon}}{numeric Station longitude where concentration (and possibly also 
discharge) was measured.}

\item{\code{basin.area}}{numeric Area of the drainage basin contributing water to the
site where concentrations were measured.}

\item{\code{flow.site.name}}{character Long name of the station where flow was 
monitored, if different from where concentration was monitored (as in 
\code{site.name}).}

\item{\code{flow.site.id}}{character Station ID of the station where flow was 
monitored, if different from where concentration was monitored (as in 
\code{site.id}).}

\item{\code{flow.lat}}{numeric Latitude of the station where flow was monitored, if 
different from where concentration was monitored (as in \code{site.lat}).}

\item{\code{flow.lon}}{numeric Longitude of the station where flow was monitored, if 
different from where concentration was monitored (as in \code{site.lon}).}

\item{\code{flow.basin.area}}{numeric Area of the drainage basin contributing water 
to the site where flow was monitored, if different from where concentration
was monitored (as in \code{basin.area}).}

\item{\code{basin.area.units}}{character Units of the values in \code{basin.area} and
\code{flow.basin.area} (must be the same for both). Basin areas are used to
compute yields as loads per contributing (basin) area.}

\item{\code{custom}}{ANY. Empty by default, but may be modified to store any 
additional data the user wants to track.}
}}

