% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata-getters}
\alias{metadata-getters}
\alias{getCol}
\alias{getUnits}
\alias{getInfo}
\title{Access information about a load model.}
\usage{
getCol(
  metadata,
  data,
  field = c("conc", "flow", "flux rate", "date"),
  attach.units = FALSE
)

getUnits(
  metadata,
  field = c("conc", "flow", "flux", "flux rate"),
  format = NA,
  stop.on.empty = FALSE
)

getInfo(
  metadata,
  field = c("constituent", "consti.name", "flow", "load.rate", "dates", "conc.units",
    "flow.units", "load.units", "load.rate.units", "station", "site.name", "site.id",
    "lat", "lon", "basin.area", "flow.site.name", "flow.site.id", "flow.lat", "flow.lon",
    "flow.basin.area", "basin.area.units", "custom"),
  stop.on.empty = FALSE
)
}
\arguments{
\item{metadata}{a metadata object from which the information will be 
retrieved}

\item{data}{a data.frame from which the data column will be retrieved}

\item{field}{character string identifying the field to extract. Partial 
matching is allowed.}

\item{attach.units}{logical. If TRUE, the return value will be a
\pkg{\link{unitted}} vector.}

\item{format}{character. The format in which units should be returned. Set to
'rloadest' to output units in a format recognized by \code{rloadest}, or
'EGRET' for crosstalk with that package, or NA to keep units as stored in
'loadflex'.}

\item{stop.on.empty}{logical. Stops f the field is empty or non-character}
}
\value{
\code{getCol} returns the specified column of data as a vector.

\code{getUnits} returns the specified units as a character string.

\code{getInfo} returns the miscellaneous information specified by
  \code{field}.
}
\description{
\code{getCol} locates a column of data specified by a field name (conc, flow,
flux rate, or date).

\code{getUnits} finds the units associated with a data type (conc, flow, 
flux, or flux rate).

\code{getInfo} returns miscellaneous information (station, custom) about a 
model.
}
\details{
These metadata-getters provide a weak wall of abstraction between metadata 
and clients of the metadata; more importantly, they do the error checking to 
make sure that the designated field or data column exists and return an 
informative error if it does not.
}
\examples{
md <- metadata(constituent="NO3", flow="DISCHARGE", 
  dates="DATE", conc.units="mg L^-1", flow.units="cfs", load.units="kg", 
  load.rate.units="kg d^-1", site.name="Lamprey River, NH")
data(lamprey_nitrate)
head(getCol(md, lamprey_nitrate, 'date'))
head(getCol(md, lamprey_nitrate, 'conc'))
md <- metadata(constituent="NO3", flow="DISCHARGE", 
  dates="DATE", conc.units="mg L^-1", flow.units="cfs", load.units="kg", 
  load.rate.units="kg d^-1", site.name="Lamprey River, NH")
getUnits(md, 'flow')
md <- metadata(constituent="NO3", flow="DISCHARGE", 
  dates="DATE", conc.units="mg L^-1", flow.units="cfs", load.units="kg", 
  load.rate.units="kg d^-1", site.name="Lamprey River, NH")
getInfo(md, 'site.name')
}
\seealso{
Other metadata:
\code{\link[=metadata]{metadata()}}
}
\concept{metadata}
