% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{updateMetadata}
\title{Create or modify the metadata for a load model.}
\usage{
metadata(
  constituent,
  flow,
  load.rate = "",
  dates,
  conc.units,
  flow.units,
  load.units,
  load.rate.units,
  station = "",
  custom = NULL,
  validate = TRUE,
  consti.name = "",
  site.name = "",
  site.id = "",
  lat = NA,
  lon = NA,
  basin.area = NA,
  flow.site.name = site.name,
  flow.site.id = site.id,
  flow.lat = lat,
  flow.lon = lon,
  flow.basin.area = basin.area,
  basin.area.units = "km^2"
)

updateMetadata(metadata, new.metadata = NA, ..., validate = TRUE)
}
\arguments{
\item{constituent}{character. The name of the solute or suspended material 
whose load is to be modeled. Also the name of the data.frame column 
describing that solute or material.}

\item{flow}{character. The name of the data.frame column describing flow 
(discharge)}

\item{load.rate}{character. The name of the data.frame column, if it exists, 
describing instantaneous load}

\item{dates}{character. The name of the data.frame column describing date 
and/or time}

\item{conc.units}{character. The units in which concentration is both input 
and exported.}

\item{flow.units}{character. The units in which flow is both input and 
exported.}

\item{load.units}{character. The units in which load (flux) is both input and
exported.}

\item{load.rate.units}{character. The units in which load (flux) is reported 
by predict.loadModel().}

\item{station}{deprecated. character. A description of the sampling station 
or site. It's now recommended to use \code{site.name} and/or \code{site.id}
instead.}

\item{custom}{ANY. Empty by default, but may be modified to store any 
additional data the user wants to track.}

\item{validate}{logical. If TRUE, validObject() must pass for the object to 
return.}

\item{consti.name}{character. Consituent long name, for use in plots and 
reports}

\item{site.name}{character. A description of the station or site where the 
concentration of the constituent was measured. This argument or
\code{site.id} is a better place for a site identifier than the
\code{station} argument, which is deprecated.}

\item{site.id}{character Station ID for the site where the concentration of 
the constituent was measured.}

\item{lat}{numeric Station latitude where concentration (and possibly also 
discharge) was measured.}

\item{lon}{numeric Station longitude where concentration (and possibly also 
discharge) was measured.}

\item{basin.area}{numeric Area of the drainage basin contributing water to 
the site where concentrations were measured.}

\item{flow.site.name}{character Long name of the station where flow was 
monitored, if different from where concentration was monitored (as in 
\code{site.name}).}

\item{flow.site.id}{character Station ID of the station where flow was 
monitored, if different from where concentration was monitored (as in 
\code{site.id}).}

\item{flow.lat}{numeric Latitude of the station where flow was monitored, if 
different from where concentration was monitored (as in \code{site.lat}).}

\item{flow.lon}{numeric Longitude of the station where flow was monitored, if
different from where concentration was monitored (as in \code{site.lon}).}

\item{flow.basin.area}{numeric Area of the drainage basin contributing water 
to the site where flow was monitored, if different from where concentration
was monitored (as in \code{basin.area}).}

\item{basin.area.units}{character Units of the values in \code{basin.area} 
and \code{flow.basin.area} (must be the same for both). Basin areas are 
used to compute yields as loads per contributing (basin) area.}

\item{metadata}{the metadata object to update.}

\item{new.metadata}{Optional object of class "metadata". If new.metadata is 
not NA, all elements of \code{...} will be ignored and metadata will be updated 
with any non-empty elements of new.metadata.}

\item{...}{named arguments describing metadata elements to be modified; the 
names may be any of those accepted by \code{metadata()}.}
}
\value{
\code{metadata} returns a new metadata object with the specified 
  entries.

\code{updateMetadata} returns a new metadata object that combines the
  values in the initial \code{metadata} with EITHER the non-empty elements of
  \code{new.metadata} OR the elements specified in \code{...}, depending on
  whether \code{new.metadata} is provided.
}
\description{
\code{metadata} creates a new metadata object (see \link{metadata-class}) 
from scratch.
}
\details{
By default, newly created metadata objects are checked for validity. This 
check may be bypassed by setting \code{validate=FALSE}. However, except where
the user is only interested in using metadata objects in isolation (i.e., 
outside a load model), this option will rarely be appropriate.

\subsection{Valid units options, \pkg{loadflex} style}{

Models used in \pkg{\link{loadflex}} require metadata with units in a 
standardized form. For any model not produced using \pkg{rloadest} functions,
this standardized format is a space-separated list of text units with 
exponents denoted with "^". Valid examples include "mg L^-1", "ft^3 s^-1", 
"kg", and "colonies d^-1". Here is the full set of possibilites, separated by
units type:

\code{flow.units} take the form VOLUME TIME^-1.

\itemize{

\item VOLUME: m^3, ft^3, dL, L

\item TIME: s, d, y

}

\code{conc.units} take the form MASS/VOLUME or COUNT/VOLUME.

\itemize{

\item MASS: ng, ug, mg, g, kg, Mg, lb, ton

\item COUNT: colonies, million_colonies

\item VOLUME: m^3, ft^3, dL, L

}

\code{load.units} take the form MASS or COUNT.

\itemize{

\item MASS: ng, ug, mg, g, kg, Mg, lb, ton

\item COUNT: colonies, million_colonies

}

\code{load.rate.units} take the form MASS/TIME or COUNT/TIME.

\itemize{

\item MASS: ng, ug, mg, g, kg, Mg, lb, ton

\item COUNT: colonies, million_colonies

\item TIME: s, d, y

}

#' \code{basin.area.units} take the form AREA.

\itemize{

\item AREA: m^2, ha, km^2, ft^2, ac, mi^2

}

For compatibility with \pkg{rloadest} models, \code{metadata} and 
\code{updateMetadata} also accept all those units accepted by \pkg{rloadest} 
(see next section).

}

\subsection{Valid units options, \pkg{rloadest} style}{

Units passed to \pkg{rloadest} take the form "X/Y" or "X per Y" or an 
abbreviation and accept a limited number of numerators and denominators. Any 
valid \pkg{rloadest} units will be translated to valid \pkg{\link{loadflex}} 
units when supplied to \code{\link{loadReg2}}, \code{metadata} or 
\code{updateMetadata}. These rloadest-style units are currently accepted:

\code{flow.units} take the form VOLUME/TIME, VOLUME per TIME, or ABV.

\itemize{

\item VOLUME: cubic meter, cubic meters, m^3, cubic foot, cubic feet, ft^3, 
100mL, dL, liter, l, L

\item TIME: second, sec, day, d, year, yr, y

\item ABV: cms, cfs

}

\code{conc.units} take the form MASS/VOLUME or COUNT/VOLUME.

\itemize{

\item MASS: nanograms, micrograms, milligrams, grams, kilograms, metric tons,
ng, ug, mg, g, kg, Mg, tons, pounds, lbs, lb

\item COUNT: col, colonies, million colonies

\item VOLUME: cubic meter, cubic meters, m^3, cubic foot, cubic feet, ft^3, 
100mL, dL, liter, l, L

}

\code{load.units} take the form MASS or COUNT.

\itemize{

\item MASS: nanograms, micrograms, milligrams, grams, kilograms, metric tons,
ng, ug, mg, g, kg, Mg, tons, pounds, lbs, lb

\item COUNT: col, colonies, million colonies

}

\code{load.rate.units} are derived as \code{load.units} per day.

}
}
\examples{
md <- metadata(constituent="NO3", flow="DISCHARGE", 
  dates="DATE", conc.units="mg L^-1", flow.units="cfs", load.units="kg", 
  load.rate.units="kg d^-1", site.name="Lamprey River, NH",
  site.id="1073500", custom=list(data_source="USGS NWIS, waterdata.usgs.gov"))
md1 <- metadata(constituent="NO3", flow="DISCHARGE", 
  dates="DATE", conc.units="mg L^-1", flow.units="cfs", load.units="kg", 
  load.rate.units="kg d^-1", site.name="Lamprey River, NH")
md2 <- updateMetadata(md1, site.id="1073500", 
  custom=list(data_source="USGS NWIS, waterdata.usgs.gov"))
}
\seealso{
Other metadata: 
\code{\link{metadata-getters}}
}
\concept{metadata}
