% Generated by roxygen2 (4.0.2): do not edit by hand
\name{metadata}
\alias{exampleMetadata}
\alias{metadata}
\alias{updateMetadata}
\title{Create or modify the metadata for a load model.}
\usage{
metadata(constituent, flow, load.rate = "", dates, conc.units, flow.units,
  load.units, load.rate.units, station = "", custom = NULL,
  validate = TRUE)

updateMetadata(metadata, new.metadata = NA, ..., validate = TRUE)

exampleMetadata()
}
\arguments{
\item{flow}{character. The name of the data.frame column describing flow
(discharge)}

\item{load.rate}{character. The name of the data.frame column, if it exists,
describing instantaneous load}

\item{dates}{character. The name of the data.frame column describing date
and/or time}

\item{conc.units}{character. The units in which concentration is both input
and exported.}

\item{flow.units}{character. The units in which flow is both input and
exported.}

\item{load.units}{character. The units in which load (flux) is both input and
exported.}

\item{load.rate.units}{character. The units in which load (flux) is reported
by predict.loadModel().}

\item{station}{character. A description of the sampling station or site.}

\item{custom}{ANY. Empty by default, but may be modified to store any
additional data the user wants to track.}

\item{validate}{logical. If TRUE, validObject() must pass for the object to
return.}

\item{new.metadata}{Optional object of class "metadata". If new.metadata is
not NA, all elements of \code{...} will be ignored and metadata will be updated
with any non-empty elements of new.metadata.}

\item{...}{named arguments describing metadata elements to be modified; the
names may be any of those accepted by \code{metadata()}.}

\item{constitutent}{character. The name of the solute or suspended material
whose load is to be modeled. Also the name of the data.frame column
describing that solute or material.}
}
\value{
\code{metadata} returns a new metadata object with the specified
  entries.

\code{updateMetadata} returns a new metadata object that combines the
  values in the initial \code{metadata} with EITHER the non-empty elements of
  \code{new.metadata} OR the elements specified in \code{...}, depending on
  whether \code{new.metadata} is provided.

\code{exampleMetadata} returns a metadata object with example entries.
}
\description{
\code{metadata} creates a new metadata object (see \link{metadata-class})
from scratch.

\code{updateMetadata} modifies an existing metadata object.

\code{exampleMetadata} produces an example metadata object, useful in testing
or to see how a complete metadata object looks.
}
\details{
By default, newly created metadata objects are checked for validity. This
check may be bypassed by setting \code{validate=FALSE}. However, except where
the user is only interested in using metadata objects in isolation (i.e.,
outside a load model), this option will rarely be appropriate.

\subsection{Valid units options, \pkg{loadflex} style}{

Models used in \pkg{\link{loadflex}} require metadata with units in a
standardized form. For any model not produced using \pkg{rloadest} functions,
this standardized format is a space-separated list of text units with
exponents denoted with "^". Valid examples include "mg L^-1", "ft^3 s^-1",
"kg", and "colonies d^-1". Here is the full set of possibilites, separated by
units type:

\code{flow.units} take the form VOLUME TIME^-1.

\itemize{

\item VOLUME: m^3, ft^3, dL, L

\item TIME: s, d

}

\code{conc.units} take the form MASS/VOLUME or COUNT/VOLUME.

\itemize{

\item MASS: ng, ug, mg, g, kg, Mg, lb, ton

\item COUNT: colonies, million_colonies

\item VOLUME: m^3, ft^3, dL, L

}

\code{load.units} take the form MASS or COUNT.

\itemize{

\item MASS: ng, ug, mg, g, kg, Mg, lb, ton

\item COUNT: colonies, million_colonies

}

\code{load.rate.units} take the form MASS/TIME or COUNT/TIME.

\itemize{

\item MASS: ng, ug, mg, g, kg, Mg, lb, ton

\item COUNT: colonies, million_colonies

\item TIME: s, d

}

For compatibility with \pkg{rloadest} models, \code{metadata} and
\code{updateMetadata} also accept all those units accepted by \pkg{rloadest}
(see next section).

}

\subsection{Valid units options, \pkg{rloadest} style}{

Units passed to \pkg{rloadest} take the form "X/Y" or "X per Y" or an
abbreviation and accept a limited number of numerators and denominators. Any
valid \pkg{rloadest} units will be translated to valid \pkg{\link{loadflex}}
units when supplied to \code{\link{loadReg2}}, \code{metadata} or
\code{updateMetadata}. These rloadest-style units are currently accepted:

\code{flow.units} take the form VOLUME/TIME, VOLUME per TIME, or ABV.

\itemize{

\item VOLUME: cubic meter, cubic meters, m^3, cubic foot, cubic feet, ft^3,
100mL, dL, liter, l, L

\item TIME: second, sec, day, d

\item ABV: cms, cfs

}

\code{conc.units} take the form MASS/VOLUME or COUNT/VOLUME.

\itemize{

\item MASS: nanograms, micrograms, milligrams, grams, kilograms, metric tons,
ng, ug, mg, g, kg, Mg, tons, pounds, lbs, lb

\item COUNT: col, colonies, million colonies

\item VOLUME: cubic meter, cubic meters, m^3, cubic foot, cubic feet, ft^3,
100mL, dL, liter, l, L

}

\code{load.units} take the form MASS or COUNT.

\itemize{

\item MASS: nanograms, micrograms, milligrams, grams, kilograms, metric tons,
ng, ug, mg, g, kg, Mg, tons, pounds, lbs, lb

\item COUNT: col, colonies, million colonies

}

\code{load.rate.units} are derived as \code{load.units} per day.

}
}
\seealso{
Other metadata: \code{\link{equals.metadata}};
  \code{\link{getCol}}, \code{\link{getInfo}},
  \code{\link{getUnits}}, \code{\link{metadata-getters}}
}

