% Generated by roxygen2 (4.0.2): do not edit by hand
\name{observeSolute}
\alias{observeSolute}
\title{observeSolute - instantaneous loads or concentrations}
\usage{
observeSolute(data, flux.or.conc = c("flux", "conc"), metadata,
  calculate = isTRUE(flux.or.conc == "flux"), attach.units = FALSE)
}
\arguments{
\item{data}{data.frame containing, at a minimum, the columns named by
metadata@constituent and metadata@flow}

\item{flux.or.conc}{character giving the desired output format}

\item{metadata}{An object of class "metadata" describing the units of flow
(flow.units) and concentration (conc.units) of the input data, and the
desired units of load (load.rate.units) for the output data}

\item{calculate}{logical. If FALSE, looks for a column containing the output
of interest. If true, uses the other two columns (out of those for conc,
flow, and flux) to calculate the output of interest.}
}
\description{
Calculates observed instantaneous loading rates or concentrations from
observed concentrations, flows, and/or fluxes, with units conversions
according to the supplied metadata.
}

