% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{observeSolute}
\alias{observeSolute}
\title{observeSolute - instantaneous loads or concentrations}
\usage{
observeSolute(data, flux.or.conc = c("flux", "conc"), metadata,
  calculate = isTRUE(flux.or.conc == "flux"), attach.units = FALSE)
}
\arguments{
\item{data}{data.frame containing, at a minimum, the columns named by
metadata@constituent and metadata@flow}

\item{flux.or.conc}{character giving the desired output format}

\item{metadata}{An object of class "metadata" describing the units of flow
(flow.units) and concentration (conc.units) of the input data, and the
desired units of load (load.rate.units) for the output data}

\item{calculate}{logical. If FALSE, looks for a column containing the output
of interest. If true, uses the other two columns (out of those for conc,
flow, and flux) to calculate the output of interest.}

\item{attach.units}{logical. If TRUE, the converted observations are returned
with units attached.}
}
\description{
Calculates observed instantaneous loading rates or concentrations from
observed concentrations, flows, and/or fluxes, with units conversions
according to the supplied metadata.
}
\examples{
obs <- data.frame(MyConc=(1:10)/10, MyFlow=rep(10,10), MyFlux=2) # intentionally inconsistent
# between conc*flow and flux
md <- updateMetadata(exampleMetadata(), constituent="MyConc", flow="MyFlow",
load.rate="MyFlux", dates="none", flow.units="cms", conc.units="mg/l",
load.units="g", load.rate.units="g/s", custom=NULL)

observeSolute(obs, "flux", md, attach.units=TRUE) # calculate flux from conc & flow
observeSolute(obs, "flux", md, calculate=FALSE, attach.units=TRUE) # read flux from data column
observeSolute(obs, "conc", md, calculate=TRUE, attach.units=TRUE) # calculate conc
# from flow & flux
}
\keyword{units}

