% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCM}
\alias{plotCM}
\title{Create plots for examining the results from the composite method}
\usage{
plotCM(flux.or.conc, show.observations, load.model, finalloads,
  observations = NULL, composite = TRUE, linear.interpolation = FALSE,
  regression = TRUE, xrange = "none", dateField = "Date",
  verbose = FALSE)
}
\arguments{
\item{flux.or.conc}{character indicating whether the plots are of 
concentrations or loads, this is used for string concatation to name the
columns of data produced through composite method.}

\item{show.observations}{true or false display observations or not.}

\item{load.model}{use this to pull names of date column and constituent name}

\item{finalloads}{these are the loads or concentrations to plot}

\item{observations}{these are the observations to plot, if you are plotting
observations}

\item{composite, }{TRUE or FALSE flag for whether or not to display the
composite results}

\item{linear.interpolation}{TRUE or FALSE flag for whether or not to display
linear interpolation of the observations}

\item{regression}{TRUE or FALSE flag for whether or not to display the linear
regression results}

\item{xrange, }{a user provded range for the xvalues (dates)}

\item{verbose}{flag for to print for debugging}
}
\description{
Not sure we want to include this function in the official API, so keeping it
internal for now.
}
\keyword{internal}

