% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEGRET.R
\name{plotEGRET}
\alias{plotEGRET}
\title{Create an EGRET-style plot
 
Use a loadflex object, but create an EGRET plot.}
\usage{
plotEGRET(plot.name, intdat = NULL, estdat = NULL, preds = NULL,
  meta = NULL, ...)
}
\arguments{
\item{plot.name}{the name of the plot the user wants to create. Current
options: ConcTime, ConcQ, FluxTimeDaily, and FluxQ. Each require different
data, see Details.}

\item{intdat}{data.frame of interpolation data}

\item{estdat}{data.frame of estimation data}

\item{preds}{data.frame of load predictions}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}

\item{...}{additional arguments to pass to the plot}
}
\description{
Create an EGRET-style plot
 
Use a loadflex object, but create an EGRET plot.
}
\details{
ConcTime, ConcQ, and FluxQ require \code{intdat, meta}. FluxTimeDaily
requires \code{intdat, estdat, preds, meta}.
}

