% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEGRET.R
\name{plotEGRET}
\alias{plotEGRET}
\title{Create an EGRET-style plot}
\usage{
plotEGRET(plot.name, intdat = NULL, estdat = NULL, preds = NULL,
  meta = NULL, preds.type = "Conc", ...)
}
\arguments{
\item{plot.name}{the name of the plot the user wants to create. See 
Details for current options. For now, only one allowed at time.}

\item{intdat}{data.frame of interpolation data}

\item{estdat}{data.frame of estimation data}

\item{preds}{data.frame of load predictions}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}

\item{preds.type}{character specifying if the predictions being used are
concentrations ("Conc") or fluxes ("Flux").}

\item{...}{additional arguments to pass to the plot}
}
\description{
Use loadflex data frames, but create an EGRET plot.
}
\details{
EGRET plots that require \code{intdat, meta}:
\itemize{
  \item boxConcMonth
  \item plotConcTime
  \item plotConcQ
  \item plotFluxQ
}
EGRET plots that require \code{intdat, estdat, preds, meta}:
\itemize{
  \item boxQTwice
  \item multiPlotDataOverview
  \item plotConcTimeDaily
  \item plotFluxTimeDaily
  \item plotConcPred 
  \item plotFluxPred 
  \item plotResidPred
  \item plotResidQ  
  \item plotResidTime
  \item boxResidMonth
  \item boxConcThree 
  \item plotConcHist 
  \item plotFluxHist 
}
}

