% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEGRET.R
\name{plotEGRET}
\alias{plotEGRET}
\title{Create an EGRET-style plot
 
Use a loadflex object, but create an EGRET plot.}
\usage{
plotEGRET(intdat, estdat, preds, meta, plotName, ...)
}
\arguments{
\item{intdat}{data.frame of interpolation data}

\item{estdat}{data.frame of estimation data}

\item{preds}{data.frame of load predictions}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: staAbbr)}

\item{plotName}{the name of the plot the user wants to create. Current
options: ConcTime, ConcQ, FluxTimeDaily, and FluxQ.}

\item{...}{additional arguments to pass to the plot}
}
\description{
Create an EGRET-style plot
 
Use a loadflex object, but create an EGRET plot.
}

