% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEGRET.R
\name{plotEGRET}
\alias{plotEGRET}
\title{Create an EGRET-style plot}
\usage{
plotEGRET(plot.name, load.model = NULL, newdata = NULL, data = NULL,
  meta = NULL, moreTitle = "loadflex", plotFlowNorm = FALSE, ...)
}
\arguments{
\item{plot.name}{the name of the plot the user wants to create. See 
Details for current options. For now, only one allowed at time.}

\item{load.model}{a load model (loadReg2, loadComp, loadInterp, loadLm, etc.)
whose data and predictions are to be converted to EGRET format}

\item{newdata}{data.frame of data used to generate predictions from an
already-fitted model}

\item{data}{data.frame of data used to fit a model. only required if 
load.model is omitted}

\item{meta}{loadflex metadata object; it must include constituent, flow, 
dates, conc.units, site.id, and consti.name. only required if load.model is
omitted}

\item{moreTitle}{additional text to include in the fluxBiasMulti plot 
title. The EGRET default is "WRTDS", so this changes the default to "loadflex".}

\item{plotFlowNorm}{logical indicating whether or not to plot the normalized flow
lines. This defaults to FALSE, which overrides the EGRET default TRUE. Applicable 
in plotFluxHist and plotConcHist.}

\item{...}{additional arguments to pass to the plot}
}
\description{
Use loadflex data frames, but create an EGRET plot.
}
\details{
EGRET plots that require \code{data, meta}:
\itemize{
  \item boxConcMonth
  \item plotConcTime
  \item plotConcQ
  \item plotFluxQ
}
EGRET plots that require \code{data, newdata, preds, meta}:
\itemize{
  \item boxQTwice
  \item multiPlotDataOverview
  \item plotConcTimeDaily
  \item plotFluxTimeDaily
  \item plotConcPred 
  \item plotFluxPred 
  \item plotResidPred
  \item plotResidQ  
  \item plotResidTime
  \item boxResidMonth
  \item boxConcThree 
  \item plotConcHist 
  \item plotFluxHist
  \item fluxBiasMulti 
}
}
\examples{

# Load necessary data + create the appropriate metadata
data(lamprey_nitrate) # interpolation data (grab sample obs)
fitdat <- lamprey_nitrate

data("lamprey_discharge")
estdat <- lamprey_discharge # estimation data (15 min interval)
estdat <- subset(estdat, DATE < as.POSIXct("2012-10-01 00:00:00", tz="EST5EDT")) 
estdat <- estdat[seq(1, nrow(estdat), by=96/4),] # only keep 4 observations per day

meta <- metadata(
  constituent="NO3", 
  flow="DISCHARGE",
  dates="DATE", 
  conc.units="mg L^-1", 
  flow.units="cfs", 
  load.units="kg",
  load.rate.units="kg d^-1", 
  site.name="Lamprey River, NH",
  site.id="01073500",
  consti.name="Nitrate")

# Run your model and get your predictions
no3_lm <- loadLm(
  formula=log(NO3) ~ log(DISCHARGE), pred.format="conc", 
  data=fitdat, metadata=meta, retrans=exp)

# Now you can plot
plotEGRET("boxConcMonth", data = lamprey_nitrate, meta = meta)
plotEGRET("multiPlotDataOverview", load.model=no3_lm, newdata=estdat)

}
