% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEGRET.R
\name{plotEGRET}
\alias{plotEGRET}
\title{Create an EGRET-style plot}
\usage{
plotEGRET(plot.name, intdat = NULL, estdat = NULL, preds = NULL,
  meta = NULL, preds.type = "Conc", ...)
}
\arguments{
\item{plot.name}{the name of the plot the user wants to create. See 
Details for current options. For now, only one allowed at time.}

\item{intdat}{data.frame of interpolation data}

\item{estdat}{data.frame of estimation data}

\item{preds}{data.frame of load predictions}

\item{meta}{loadflex metadata object; it must include constituent,
flow, conc.units, custom (station abbreviation: sta.abbr, and a short
name for the constituent: consti.name)}

\item{preds.type}{character specifying if the predictions being used are
concentrations ("Conc") or fluxes ("Flux").}

\item{...}{additional arguments to pass to the plot}
}
\description{
Use loadflex data frames, but create an EGRET plot.
}
\details{
EGRET plots that require \code{intdat, meta}:
\itemize{
  \item boxConcMonth
  \item plotConcTime
  \item plotConcQ
  \item plotFluxQ
}
EGRET plots that require \code{intdat, estdat, preds, meta}:
\itemize{
  \item boxQTwice
  \item multiPlotDataOverview
  \item plotConcTimeDaily
  \item plotFluxTimeDaily
  \item plotConcPred 
  \item plotFluxPred 
  \item plotResidPred
  \item plotResidQ  
  \item plotResidTime
  \item boxResidMonth
  \item boxConcThree 
  \item plotConcHist 
  \item plotFluxHist 
}
}
\examples{

# Load necessary data + create the appropriate metadata
data(lamprey_nitrate) # interpolation data (grab sample obs)

data("lamprey_discharge")
estdat <- lamprey_discharge # estimation data (15 min interval)
estdat <- subset(estdat, DATE < as.POSIXct("2012-10-01 00:00:00", tz="EST5EDT")) 
estdat <- estdat[seq(1, nrow(estdat), by=96/4),] # only keep 4 observations per day

meta <- metadata(constituent="NO3", 
                 flow="DISCHARGE",
                 dates="DATE", 
                 conc.units="mg L^-1", 
                 flow.units="cfs", 
                 load.units="kg",
                 load.rate.units="kg d^-1", 
                 station="Lamprey River, NH",
                 custom=list(sta.abbr = "lamp",
                             consti.name = "nitrate"))

# Run your model and get your predictions
no3_lm <- loadLm(formula=log(NO3) ~ log(DISCHARGE), pred.format="conc", 
                 data=intdat, metadata=meta, retrans=exp)
preds <- predictSolute(no3_lm, "conc", estdat, se.pred=TRUE, date=TRUE)

# Now you can plot
plotEGRET("boxConcMonth", intdat = lamprey_nitrate, meta = meta)
plotEGRET("multiPlotDataOverview", lamprey_nitrate, estdat, preds, meta)

}

