% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{plotResidualsCM}
\alias{plotResidualsCM}
\title{plot residuals}
\usage{
plotResidualsCM(type, load.model, observations, dateField = "Date",
  verbose = FALSE, xObservations = FALSE, xFlow = FALSE,
  day.of.year = FALSE)
}
\arguments{
\item{load.model}{use this to pull names of date column and constituent name}

\item{observations}{these are the observations to plot the residuals for}

\item{dateField}{name of the date field in observations}

\item{verbose}{debugging flag}

\item{day.of.year}{flag if true will plot residuals by the day of year to, can help user look for senasonality}
}
\description{
plot residuals for visual analysis.
}

