% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadComp.R
\name{predictSolute.loadComp}
\alias{predictSolute.loadComp}
\title{Make flux or concentration predictions from a loadComp model.}
\usage{
\method{predictSolute}{loadComp}(load.model, flux.or.conc, newdata,
  interval = c("none", "confidence", "prediction"), level = 0.95,
  se.fit = FALSE, se.pred = FALSE, date = FALSE, attach.units = FALSE,
  fit.reg = FALSE, fit.resid = FALSE, fit.resid.raw = FALSE, ...)
}
\arguments{
\item{load.model}{A loadComp object.}

\item{flux.or.conc}{indicate if the solute is measured as a flux or 
concentration.}

\item{newdata}{\code{data.frame}, optional. Predictor data, including any 
predictor columns needed for the regression model and any needed for the 
interpolation model. Column names should match those given in the 
\code{loadComp} metadata. If \code{newdata} is not supplied, the original 
fitting data will be used.}

\item{interval}{character. One of "none", "confidence" or "prediction".
If "confidence" or "prediction", the interval bounds will be returned in 
columns named "lwr" and "upr". Confidence intervals describe confidence in 
the model prediction for the mean value given a set of predictors, whereas 
prediction bounds describe the expected distribution of observations at 
that prediction point.}

\item{level}{numeric. Fraction of density distribution to include within confidence or
prediction interval}

\item{se.fit}{logical. If TRUE, the output data.frame will include a column 
named "se.fit" describing the standard error of the model fit for each row 
of predictors.}

\item{se.pred}{logical. If TRUE, the output data.frame will include a column 
named "se.pred" describing the standard error of the prediction for each 
row of predictors. The values in the se.pred column will be larger than 
those in the se.fit column, because the se.pred values are standard errors 
of prediction (SEPs) and take into account not only the parameter 
uncertainty associated with the model coefficients (also covered by 
se.fit), but also the random error associated with any given observation 
(the epsilon term in a typical regression model).}

\item{date}{logical. If TRUE, the output data.frame will include a column 
named "date" containing the dates of the predictions.}

\item{attach.units}{logical. Should the units be attached to columns in the 
resulting data.frame?}

\item{fit.reg}{logical. If TRUE, a column for the regression predictions 
before interpolation will be included in the data.frame that is returned. 
These will be in the same format (non-log, conc/flux) as the final 
predictions.}

\item{fit.resid}{logical. If TRUE, a column for the residual corrections will
be included in the data.frame that is returned. These will be in the same 
format (non-log, conc/flux) as the final predictions, even when the 
residuals were actually produced in log space and/or as relative residuals,
so that fit = fit.reg + fit.resid.}

\item{fit.resid.raw}{logical. If TRUE, a column for the residual corrections 
as returned from the interpolation model. These may be in log space and/or 
unitless, depending on the type of residual correction specified when the
loadComp model was created.}

\item{...}{Additional arguments passed to class-specific implementations of 
the \code{predictSolute} generic function.}
}
\value{
A vector of data.frame of predictions, as for the generic 
  \code{\link{predictSolute}}.
}
\description{
Makes instantaneous predictions (at the temporal resolution of 
\code{newdata}) from a fitted \code{\link{loadComp}} model. See 
\code{\link{predictSolute}} for details.
}
\seealso{
Other predictSolute: \code{\link{predictSolute.loadInterp}},
  \code{\link{predictSolute.loadLm}},
  \code{\link{predictSolute.loadModel}},
  \code{\link{predictSolute.loadReg2}},
  \code{\link{predictSolute}}
}
