% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLm.R
\name{predictSolute.loadLm}
\alias{predictSolute.loadLm}
\title{Make flux or concentration predictions from a loadLm model.}
\usage{
\method{predictSolute}{loadLm}(load.model, flux.or.conc = c("flux", "conc"),
  newdata, interval = c("none", "confidence", "prediction"), level = 0.95,
  lin.or.log = c("linear", "log"), se.fit = FALSE, se.pred = FALSE,
  date = FALSE, attach.units = FALSE, agg.by = c("unit", "day", "month",
  "water year", "calendar year", "total", "mean water year",
  "mean calendar year", "[custom]"), ...)
}
\arguments{
\item{load.model}{A loadLm object.}

\item{flux.or.conc}{character. Should the predictions be reported as flux
rates or concentrations? If the output is a data.frame, the column name for
flux predictions will be "fit" when `agg.by='unit'` and "Flux_Rate"
otherwise; the column name for concentration predictions will be "fit" when
`agg.by='unit'` and "Conc" otherwise.}

\item{newdata}{\code{data.frame}, optional. Predictor data. Column names
should match those given in the \code{loadLm} metadata. If
\code{newdata} is not supplied, the original fitting data will be used.}

\item{interval}{character. One of "none", "confidence" or "prediction". If
"confidence" or "prediction", the interval bounds will be returned in
columns named "lwr" and "upr". Confidence intervals describe confidence in
the model prediction for the mean value given a set of predictors, whereas
prediction bounds describe the expected distribution of observations at
that prediction point.}

\item{level}{numeric. Fraction of density distribution to include within
confidence or prediction interval}

\item{lin.or.log}{character. Either "linear" or "log" to say whether the
predictions should be converted to log space or not. If converted to log
space, a bias correction will be applied to regression model predictions;
see \code{\link{linToLog}}.}

\item{se.fit}{logical. If TRUE, the output data.frame will include a column
named "se.fit" (for agg.by=="unit") or "SE" (for agg.by!="unit") describing
the standard error of the model fit for each row of predictors.}

\item{se.pred}{logical. If TRUE, the output data.frame will include a column
named "se.pred" (for agg.by=="unit") or "SE" (for agg.by!="unit")
describing the standard error of the prediction for each row of predictors.
Only one of se.fit or se.pred is permitted for agg.by!="unit". The se.pred
values are standard errors of prediction (SEPs) and take into account not
only the parameter uncertainty associated with the model coefficients (also
covered by se.fit), but also the random error associated with any given
observation (the epsilon term in a typical regression model).}

\item{date}{logical. If TRUE, the output data.frame will include a column
named "date" containing the dates of the predictions (for agg.by=="unit"),
or a period description such as "Water_Year" if the predictions have been
aggregated (for agg.by!="unit").}

\item{attach.units}{logical. Should the units be attached to columns in the
resulting data.frame?}

\item{agg.by}{character Time period to aggregate results by. To do no
aggregation, use the default of `agg.by='unit'`.}

\item{...}{Additional arguments passed to class-specific implementations of
the \code{predictSolute} generic function.}
}
\value{
A vector of data.frame of predictions, as for the generic 
  \code{\link{predictSolute}}.
}
\description{
Makes instantaneous predictions (at the temporal resolution of 
\code{newdata}) from a fitted \code{\link{loadLm}} model. See 
\code{\link{predictSolute}} for details.
}
\seealso{
Other predictSolute: \code{\link{predictSolute.loadComp}},
  \code{\link{predictSolute.loadInterp}},
  \code{\link{predictSolute.loadModel}},
  \code{\link{predictSolute.loadReg2}},
  \code{\link{predictSolute}}
}
