% Generated by roxygen2 (4.0.2): do not edit by hand
\name{residDurbinWatson}
\alias{residDurbinWatson}
\title{Test for autocorrelation of residuals}
\usage{
residDurbinWatson(load.model, flux.or.conc = c("flux", "conc"),
  abs.or.rel.resids = c("absolute", "relative"), newdata = NULL,
  plot.acf = TRUE, timestep.tol = .Machine$double.eps^0.5,
  irregular.timesteps.ok = FALSE)
}
\arguments{
\item{load.model}{a loadModel descendant}

\item{flux.or.conc}{character. The format in which residuals should be
calculated}

\item{abs.or.rel.resids}{character. Should residuals be computed as the
difference or the ratio of the observed and predicted values?}

\item{newdata}{The data from which to compute residuals; if NULL, the
original fitting data for load.model will be used.}

\item{plot.acf}{logical. Should the autocorrelation function be plotted?}

\item{irregular.timesteps.ok}{logical. By default, this function requires
  that the timesteps between observations are identical to one another, and
  an error is thrown if this requirement is not met. The check is not
  performed if \code{irregular.timesteps.ok} is TRUE.}
}
\value{
A Durbin-Watson test statistic applied to residuals.
}
\description{
Extracts residuals from a load.model (where residuals may be for the
calibration data or for a new set of observations). Applies
car::durbinWatsonTest to test for autocorrelation of those residuals.
}
\seealso{
car::durbinWatsonTest

Other diagnostics: \code{\link{estimateRho}};
  \code{\link{getCorrectionFraction}},
  \code{\link{getCorrectionFraction.loadComp}}
}

