% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sentenceCase}
\alias{.sentenceCase}
\alias{sentenceCase}
\title{Capitalize first letter of words.}
\usage{
.sentenceCase(x)
}
\arguments{
\item{x}{a character vector to convert}
}
\description{
Convert each word in x to have an upper-case first letter and lower-case
 subsequent letters. Doesn't make exceptions for articles like "the" or "a".
}

