% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\docType{methods}
\name{show.metadata}
\alias{show.metadata}
\title{Display a metadata object}
\usage{
object
}
\arguments{
\item{object}{The metadata object to be displayed}
}
\description{
Display a metadata object
}

