% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show.metadata}
\alias{show.metadata}
\title{Display a metadata object}
\usage{
\S4method{show}{metadata}(object)
}
\arguments{
\item{object}{The metadata object to be displayed}
}
\description{
Display a metadata object
}

