% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{methods}
\name{show,metadata-method}
\alias{show,metadata-method}
\title{Display a metadata object}
\usage{
\S4method{show}{metadata}(object)
}
\arguments{
\item{object}{The metadata object to be displayed}
}
\description{
Display a metadata object
}

