% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeInputs}
\alias{summarizeInputs}
\title{Summarize the site and input data}
\usage{
summarizeInputs(metadata, fitdat, estdat)
}
\arguments{
\item{metadata}{metadata, used to access the appropriate columns of data. At 
a minimum, \code{metadata} should correctly specify the date column and the
column indicated by \code{interp.format}.}

\item{fitdat}{data frame record of constituent+discharge measurements, for
fitting a model}

\item{estdat}{data frame record of discharge measurements, for making
predictions from a model}
}
\value{
data frame of the following information:
}
\description{
Summarize the site and input data
}
\examples{
data(lamprey_nitrate)
data(lamprey_discharge)
md <- metadata(constituent="NO3", flow="DISCHARGE", 
  dates="DATE", conc.units="mg L^-1", flow.units="cfs", load.units="kg", 
  load.rate.units="kg d^-1", site.name="Lamprey River, NH",
  basin.area=50, flow.basin.area=65, basin.area.units='ha',
  site.id="1073500", custom=list(data.source="USGS NWIS, waterdata.usgs.gov"))
sitesum <- summarizeInputs(metadata=md, fitdat=select(lamprey_nitrate, -REGR), 
  estdat=lamprey_discharge)
}

