% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadComp.R
\name{summarizeModel.loadComp}
\alias{summarizeModel.loadComp}
\title{Extract model summary statistics from a loadComp model}
\usage{
\method{summarizeModel}{loadComp}(load.model, ...)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel 
and always implementing the loadModelInterface.}

\item{...}{Other arguments passed to model-specific methods}
}
\value{
A 1-row data.frame of model metrics
}
\description{
Produce a 1-row data.frame of model metrics. The relevant metrics for 
loadComp models include two sets of statistics about autocorrelation (one for
the regression residuals, one for the 'residuals' used to do the composite
correction).
}
\seealso{
Other summarizeModel: \code{\link{summarizeModel.loadInterp}},
  \code{\link{summarizeModel.loadLm}},
  \code{\link{summarizeModel.loadReg}},
  \code{\link{summarizeModel}}
}

