% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInterp.R
\name{summarizeModel.loadInterp}
\alias{summarizeModel.loadInterp}
\title{Extract model summary statistics from a loadInterp model}
\usage{
\method{summarizeModel}{loadInterp}(load.model, ...)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel 
and always implementing the loadModelInterface.}

\item{...}{Other arguments passed to model-specific methods}
}
\value{
A 1-row data.frame of model metrics
}
\description{
Produce a 1-row data.frame of model metrics. The relevant metrics for 
loadInterp models include RMSE, p-values, and others TBD.
}
\seealso{
Other summarizeModel: \code{\link{summarizeModel.loadComp}},
  \code{\link{summarizeModel.loadLm}},
  \code{\link{summarizeModel.loadReg}},
  \code{\link{summarizeModel}}
}

