% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLm.R
\name{summarizeModel.loadLm}
\alias{summarizeModel.loadLm}
\title{Extract model summary statistics from a loadLm model}
\usage{
\method{summarizeModel}{loadLm}(load.model, ...)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel 
and always implementing the loadModelInterface.}

\item{...}{Other arguments passed to model-specific methods}
}
\value{
A 1-row data.frame of model metrics
}
\description{
Produce a 1-row data.frame of model metrics. The relevant metrics for 
loadLm models include RMSE, p-values, and others TBD.
}
\seealso{
Other summarizeModel: \code{\link{summarizeModel.loadComp}},
  \code{\link{summarizeModel.loadInterp}},
  \code{\link{summarizeModel.loadReg}},
  \code{\link{summarizeModel}}
}

