% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R, R/loadReg2.R
\name{summarizeModel.loadReg}
\alias{summarizeModel.loadReg}
\alias{summarizeModel.loadReg2}
\title{Extract model summary statistics from a loadReg2 model}
\usage{
\method{summarizeModel}{loadReg}(load.model, flux.or.conc = c("flux", "conc"))

\method{summarizeModel}{loadReg2}(load.model)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel 
and always implementing the loadModelInterface.}
}
\value{
A 1-row data.frame of model metrics
}
\description{
Produce a 1-row data.frame of model metrics. The relevant metrics for 
loadReg2 models are largely the same as those reported by the \code{rloadest}
package, though reported in this streamlined data.frame format for bulk 
reporting. \code{summarizeModel.loadReg} should rarely be accessed directly;
instead, call \code{summarizeModel()} on a \code{loadReg2} object.
}
\section{Metrics}{

  
  \describe{
  
  \item{\code{Intercept}, \code{lnQ}, \code{DECTIME}, \code{sin.DECTIME}, 
  \code{cos.DECTIME}, etc.}{Coefficient estimates for the named terms. The 
  model formula determines which terms are included. Recall that log(Q) was 
  centered before the coefficients were fit.}
  
  \item{\code{RMSE}}{The root mean squared error of the difference between 
  observed and predicted values of concentration.}
  
  \item{\code{r.squared}}{}
  
  \item{\code{p.value}}{}
  
  }
}
\seealso{
Other summarizeModel: \code{\link{summarizeModel}}
}

