% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReg.R, R/loadReg2.R
\name{summarizeModel.loadReg}
\alias{summarizeModel.loadReg}
\alias{summarizeModel.loadReg2}
\title{Extract model summary statistics from a loadReg2 model}
\usage{
\method{summarizeModel}{loadReg}(load.model, flux.or.conc = c("flux", "conc"),
  ...)

\method{summarizeModel}{loadReg2}(load.model, ...)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel 
and always implementing the loadModelInterface.}

\item{flux.or.conc}{character. Which internal model (the flux model or the 
concentration model) should be summarized? A \pkg{rloadest} model, and 
therefore also a \code{loadReg2} model, is actually two different models 
for (1) flux and (2) concentration, each fitted to the same data and with 
the same model structure except for whether the left-hand side of the model
formula is flux or concentration. Some of the model metrics differ between 
these two internal models.}

\item{...}{Other arguments passed to model-specific methods}

\item{constituent}{char the constituent being predicted}

\item{site}{char Site ID water quality measurements}
}
\value{
A 1-row data.frame of model metrics
}
\description{
Produce a 1-row data.frame of model metrics. The relevant metrics for 
loadReg2 models are largely the same as those reported by the \code{rloadest}
package, though reported in this streamlined data.frame format for bulk 
reporting. \code{summarizeModel.loadReg} should rarely be accessed directly;
instead, call \code{summarizeModel()} on a \code{loadReg2} object.
}
\section{Metrics}{

  
  \describe{
  
  \item{\code{Intercept}, \code{lnQ}, \code{DECTIME}, \code{sin.DECTIME}, 
  \code{cos.DECTIME}, etc.}{Coefficient estimates for the named terms. The 
  model formula determines which terms are included. Recall that log(Q) was 
  centered before the coefficients were fit.}
  
  \item{\code{RMSE}}{The root mean squared error of the difference between 
  observed and predicted values of concentration.}
  
  \item{\code{r.squared}}{The proportion of variation explained by the 
  model.}
  
  \item{\code{p.value}}{}
  
  }
}
\seealso{
Other summarizeModel: \code{\link{summarizeModel.loadComp}},
  \code{\link{summarizeModel.loadInterp}},
  \code{\link{summarizeModel.loadLm}},
  \code{\link{summarizeModel.loadModel}},
  \code{\link{summarizeModel}}
}

