% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelInterface.R
\name{summarizeModel}
\alias{summarizeModel}
\title{Extract model summary statistics}
\usage{
summarizeModel(load.model, ...)
}
\arguments{
\item{load.model}{A load model object, typically inheriting from loadModel
and always implementing the loadModelInterface.}

\item{...}{Other arguments passed to model-specific methods}
}
\description{
summarizeModel produces a 1-row data.frame of model metrics. The relevant
metrics vary by model type; only the relevant metrics are reported for each
model.
}
\seealso{
Other loadModelInterface:
\code{\link[=estimateMSE]{estimateMSE()}},
\code{\link[=getFittedModel]{getFittedModel()}},
\code{\link[=getFittingData]{getFittingData()}},
\code{\link[=getFittingFunction]{getFittingFunction()}},
\code{\link[=getMetadata]{getMetadata()}},
\code{\link[=predictSolute]{predictSolute()}},
\code{\link[=simulateSolute]{simulateSolute()}},
\code{\link[=validLoadModelInterface]{validLoadModelInterface()}}

Other summarizeModel:
\code{\link[=summarizeModel.loadComp]{summarizeModel.loadComp()}},
\code{\link[=summarizeModel.loadInterp]{summarizeModel.loadInterp()}},
\code{\link[=summarizeModel.loadLm]{summarizeModel.loadLm()}},
\code{\link[=summarizeModel.loadModel]{summarizeModel.loadModel()}},
\code{\link[=summarizeModel.loadReg2]{summarizeModel.loadReg2()}}
}
\concept{loadModelInterface}
\concept{summarizeModel}
