% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeModel}
\alias{summarizeModel}
\alias{summarizeModel.loadReg2}
\title{Extract various model summary statistics}
\usage{
summarizeModel(model)

\method{summarizeModel}{loadReg2}(model)
}
\arguments{
\item{model}{the input model object; will extend to several types}
}
\description{
Extract various model summary statistics

stats for rloadest loadReg2 model
}

