% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizePreds}
\alias{summarizePreds}
\title{Summarize generated predictions}
\usage{
summarizePreds(preds, meta, by, model.name, level = 0.95)
}
\arguments{
\item{preds}{data frame input data frame of predicted solutes}

\item{meta}{loadflex metadata object for the preds data frame}

\item{by}{character "total" to return average flux over all years, "annual" to return 
yearly (water year) averages}

\item{model.name}{char name of model used to generate predictions}

\item{level}{numeric level of confidence interval to return}
}
\value{
data frame containing various statistics
}
\description{
Summarize generated predictions
}

