% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeSites}
\alias{summarizeSites}
\title{summarize WQ site information}
\usage{
summarizeSites(sites, siteInfo, nutriDF)
}
\arguments{
\item{sites}{character ID(s) of sites to summarize}

\item{siteInfo}{data frame containing site information}

\item{nutriDF}{data frame record of nutrient measurements}
}
\value{
data frame of the following information:
}
\description{
summarize WQ site information
}

