% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{unit.conversions}
\alias{unit.conversions}
\title{Defines the units that are permissible in load models, along with the
multiplers that will allow common conversions among units}
\format{A data.frame with character columns "numerator", "denominator", and
  numeric column "value", one conversion per row}
\description{
Defines the units that are permissible in load models, along with the
multiplers that will allow common conversions among units
}
\keyword{data}
\keyword{units}

