% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{validDim}
\alias{validDim}
\title{Determine whether a 1D dimension string is known and of the expected
dimension type}
\usage{
validDim(dimstr, dim.type = c("ANY", "volume", "time", "mass", "count",
  "area"))
}
\arguments{
\item{dimstr}{A string representing one units dimension (just one at a time, 
please). For example: kg, d, or m^3, but not kg/d or d^-1}

\item{dim.type}{One or more acceptable dimension types}
}
\description{
Determine whether a 1D dimension string is known and of the expected
dimension type
}
\examples{
loadflex:::validDim('kg') # TRUE
loadflex:::validDim('kg', 'mass') # TRUE
loadflex:::validDim('kg', 'volume') # FALSE
loadflex:::validDim('whoknows') # FALSE
loadflex:::validDim('whoknows', 'time') # FALSE
}
