% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelInterface.R
\name{validLoadModelInterface}
\alias{validLoadModelInterface}
\title{Test whether a class implements the loadModelInterface}
\usage{
validLoadModelInterface(object, stop.on.error = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{an object with a LoadModelInterface}

\item{stop.on.error}{logical. If the interface is invalid, should the
function throw an error (TRUE) or quietly return a warning object (FALSE)?}

\item{verbose}{logical. turn on or off verbose messages.}
}
\value{
TRUE if interface for given load.model is well defined; otherwise,
  either throws an error (if stop.on.error=TRUE) or returns a vector of
  character strings describing the errors (if stop.on.error=FALSE).
}
\description{
\code{validLoadModelInterface} can be used to test whether the
\code{\link{loadModelInterface}} has been successfully implemented for the class of a
provided object.
}
\seealso{
Other loadModelInterface: 
\code{\link{estimateMSE}()},
\code{\link{getFittedModel}()},
\code{\link{getFittingData}()},
\code{\link{getFittingFunction}()},
\code{\link{getMetadata}()},
\code{\link{predictSolute}()},
\code{\link{simulateSolute}()},
\code{\link{summarizeModel}()}
}
\concept{loadModelInterface}
