% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unit.conversions.R
\name{validMetadataUnits}
\alias{validMetadataUnits}
\title{Check whether these units are acceptable (without translation) for inclusion in metadata}
\usage{
validMetadataUnits(unitstr, unit.type = c("ANY", "flow.units", "conc.units",
  "load.units", "load.rate.units"))
}
\arguments{
\item{unitstr}{A string representing units (just one bundle at a time, please)}

\item{type}{A string describing the type of units desired}
}
\value{
logical. TRUE if valid for that unit type, FALSE otherwise
}
\description{
Check whether these units are acceptable (without translation) for inclusion in metadata
}

