% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToEGRET.R
\name{verify_meta}
\alias{verify_meta}
\title{Verify metadata object
 
Verify that the loadflex metadata object has everything it needs.}
\usage{
verify_meta(meta, nm)
}
\arguments{
\item{meta}{loadflex metadata object}

\item{nm}{character name of the metadata item to check. If it is a 
custom name, this would be a character vector with the first name as
'custom' (e.g. nm = c('custom', 'staAbbr'))}
}
\description{
Verify metadata object
 
Verify that the loadflex metadata object has everything it needs.
}

